Further clean ups
