+++ runtime/src/parallel.cpp
@@ -1,6 +1,6 @@
 #include "octree.h"
 
-//#define USE_MPI
+#define USE_MPI
 
 #ifdef USE_MPI
 #include <xmmintrin.h>
@@ -125,16 +125,7 @@ void MPIComm_free_type()
   if (MPI_V4SF) MPI_Type_free(&MPI_V4SF);
 }
 
-
-
 #endif
-
-//#if ENABLE_LOG
-//extern bool ENABLE_RUNTIME_LOG;
-//extern bool PREPEND_RANK;
-//#endif
-
-
 
 
 inline int host_float_as_int(float val)
@@ -918,10 +909,6 @@ double get_time2() {
 
 
 
-
-
-
-
 #endif
 
 void octree::mpiInit(int argc,char *argv[], int &procId, int &nProcs)
@@ -938,15 +925,8 @@ void octree::mpiInit(int argc,char *argv[], int &procId, int &nProcs)
     MPI_Init(&argc,&argv);
     int provided;
 
-//    MPI_Init_thread(&argc, &argv, MPI_THREAD_FUNNELED , &provided);
-//    assert(provided == MPI_THREAD_FUNNELED );
-
-
     //MPI_Init_thread(&argc, &argv, MPI_THREAD_MULTIPLE, &provided);
     //assert(provided == MPI_THREAD_MULTIPLE);
-
-    //      MPI_Init_thread(&argc, &argv, MPI_THREAD_FUNNELED, &provided);
-    //      assert(provided == MPI_THREAD_FUNNELED);
   }
 
   MPI_Comm_size(mpiCommWorld, &nProcs);
@@ -959,9 +939,8 @@ void octree::mpiInit(int argc,char *argv[], int &procId, int &nProcs)
   char processor_name[] = "Default";
 #endif
 
-#ifdef PRINT_MPI_DEBUG
-  LOGF(stderr, "Proc id: %d @ %s , total processes: %d (mpiInit) \n", procId, processor_name, nProcs);
-#endif
+
+  LOGF(   stderr, "Proc id: %d @ %s , total processes: %d (mpiInit) \n", procId, processor_name, nProcs);
   fprintf(stderr, "Proc id: %d @ %s , total processes: %d (mpiInit) \n", procId, processor_name, nProcs);
 
   currentRLow  = new double4[nProcs];
@@ -999,15 +978,13 @@ void octree::AllSum(double &value)
 #endif
 }
 
-int octree::SumOnRootRank(int &value)
+int octree::SumOnRootRank(int value)
 {
+  int temp = value;
 #ifdef USE_MPI
-  int temp;
   MPI_Reduce(&value,&temp,1, MPI_INT, MPI_SUM,0, mpiCommWorld);
-  return temp;
-#else
-  return value;
 #endif
+  return temp;
 }
 //end utility
 
@@ -1019,60 +996,6 @@ int octree::SumOnRootRank(int &value)
 //Functions related to domain decomposition
 
 
-//Uses one communication by storing data in one buffer and communicate required information,
-//such as box-sizes and number of sample particles on this process (Note that the number is only
-//used by process-0
-void octree::sendCurrentRadiusAndSampleInfo(real4 &rmin, real4 &rmax, int nsample, int *nSamples)
-{
-
-  sampleRadInfo curProcState;
-
-  curProcState.nsample      = nsample;
-  curProcState.rmin         = make_double4(rmin.x, rmin.y, rmin.z, rmin.w);
-  curProcState.rmax         = make_double4(rmax.x, rmax.y, rmax.z, rmax.w);
-#ifdef USE_MPI
-  //Get the number of sample particles and the domain size information
-  MPI_Allgather(&curProcState, sizeof(sampleRadInfo), MPI_BYTE,  curSysState,
-      sizeof(sampleRadInfo), MPI_BYTE, mpiCommWorld);
-#else
-  curSysState[0] = curProcState;
-#endif
-
-  rmin.x                 = (real)(currentRLow[0].x = curSysState[0].rmin.x);
-  rmin.y                 = (real)(currentRLow[0].y = curSysState[0].rmin.y);
-  rmin.z                 = (real)(currentRLow[0].z = curSysState[0].rmin.z);
-  currentRLow[0].w = curSysState[0].rmin.w;
-
-  rmax.x                 = (real)(currentRHigh[0].x = curSysState[0].rmax.x);
-  rmax.y                 = (real)(currentRHigh[0].y = curSysState[0].rmax.y);
-  rmax.z                 = (real)(currentRHigh[0].z = curSysState[0].rmax.z);
-  currentRHigh[0].w = curSysState[0].rmax.w;
-
-  nSamples[0] = curSysState[0].nsample;
-
-  for(int i=1; i < nProcs; i++)
-  {
-    rmin.x = std::min(rmin.x, (real)curSysState[i].rmin.x);
-    rmin.y = std::min(rmin.y, (real)curSysState[i].rmin.y);
-    rmin.z = std::min(rmin.z, (real)curSysState[i].rmin.z);
-
-    rmax.x = std::max(rmax.x, (real)curSysState[i].rmax.x);
-    rmax.y = std::max(rmax.y, (real)curSysState[i].rmax.y);
-    rmax.z = std::max(rmax.z, (real)curSysState[i].rmax.z);
-
-    currentRLow[i].x = curSysState[i].rmin.x;
-    currentRLow[i].y = curSysState[i].rmin.y;
-    currentRLow[i].z = curSysState[i].rmin.z;
-    currentRLow[i].w = curSysState[i].rmin.w;
-
-    currentRHigh[i].x = curSysState[i].rmax.x;
-    currentRHigh[i].y = curSysState[i].rmax.y;
-    currentRHigh[i].z = curSysState[i].rmax.z;
-    currentRHigh[i].w = curSysState[i].rmax.w;
-
-    nSamples[i] = curSysState[i].nsample;
-  }
-}
 
 void octree::computeSampleRateSFC(float lastExecTime, int &nSamples, float &sampleRate)
 {
@@ -1493,17 +1416,14 @@ void octree::sendCurrentRadiusInfo(real4 &rmin, real4 &rmax)
 void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
 {
 #ifdef USE_MPI
-
   double tStart = get_time();
 
   uint4 lowerBoundary = boundaries[this->procId];
   uint4 upperBoundary = boundaries[this->procId+1];
 
-
   static std::vector<uint>  nParticlesPerDomain(nProcs);
   static std::vector<uint2> domainId           (nProcs);
 
-#if 1
   my_dev::dev_mem<uint2>  validList2(devContext);
   my_dev::dev_mem<uint2>  validList3(devContext);
   my_dev::dev_mem<uint4>  boundariesGPU(devContext);
@@ -1531,15 +1451,13 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
   domainCheckSFCAndAssign.set_arg<int   >(8,   &procId);
   domainCheckSFCAndAssign.setWork(localTree.n, 128);
   domainCheckSFCAndAssign.execute(execStream->s());
+  execStream->sync();
 
   //After this we don't need boundariesGPU anymore so can overwrite that memory space
   my_dev::dev_mem<uint>   outputValues(devContext);
   my_dev::dev_mem<uint2>  outputKeys  (devContext);
   tempOffset = outputValues.cmalloc_copy(localTree.generalBuffer1, nProcs, tempOffset );
   tempOffset = outputKeys  .cmalloc_copy(localTree.generalBuffer1, nProcs, tempOffset );
-
-  execStream->sync();
-
 
   double tCheck = get_time();
   uint2 res = thrust_partitionDomains(validList2, validList3,
@@ -1564,10 +1482,9 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
   double tExtract             = 0;
   double ta2aSize             = 0;
 
-
-  int *nparticles  = &exchangePartBuffer[0*(nProcs+1)]; //Size nProcs+1
-  int *nsendDispls = &exchangePartBuffer[1*(nProcs+1)]; //Size nProcs+1
-  int *nreceive    = &exchangePartBuffer[2*(nProcs+1)]; //Size nProcs
+  int *nparticles  = &exchangePartBuffer[0*(nProcs+1)]; //nParticles to send per domain
+  int *nreceive    = &exchangePartBuffer[2*(nProcs+1)]; //nParticles to receive per domain
+  int *nsendDispls = &exchangePartBuffer[1*(nProcs+1)]; //Prefix sum for storage
 
   //TODO
   // This can be changed by a copy per domain. That way we do not have to wait till everything is
@@ -1588,8 +1505,8 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
     //Thread 1, will do the MPI all2all stuff
     if(tid == 0)
     {
-      //Check if the memory size, of the generalBuffer is large enough to store the exported particles
-        //if not allocate more but make sure that the copy of compactList survives
+        //Check if the memory size, of the generalBuffer is large enough to store the exported particles
+        //if not allocate a buffer on the host that will store the data instead
         int validCount = nExportParticles;
         int tempSize   = localTree.generalBuffer1.get_size() - tempOffset1;
         int stepSize   = (tempSize / (sizeof(bodyStruct) / sizeof(int)))-512; //Available space in # of bodyStructs
@@ -1712,15 +1629,11 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
       for(int i=0; i < domainId.size(); i++)
       {
         const int domain = domainId[i].x & 0x0FFFFFF;
-
         assert(domain != procId); //Should not send to ourselves
-
 
         nparticles [domain] = nParticlesPerDomain[i];
         nsendDispls[domain] = sendOffset;
         sendOffset         += nParticlesPerDomain[i];
-
-        //LOGF(stderr,"Domain info: %d -> %d \n",  domainId[i].x & 0x0FFFFFF, nParticlesPerDomain[i]);
       }
 
       double tStarta2a = get_time();
@@ -1738,9 +1651,6 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
   this->gpu_exchange_particles_with_overflow_check_SFC2(localTree, &extraBodyBuffer[0],
                                                         nparticles, nsendDispls, nreceive,
                                                         nExportParticles);
-
-
-
   double tEnd = get_time();
 
   char buff5[1024];
@@ -1749,169 +1659,6 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
   devContext.writeLogEvent(buff5);
 
   if(!doInOneGo) delete[] extraBodyBuffer;
-
-#else
-
-
-  tStart = get_time();
-
-  //Memory buffers to hold the extracted particle information
-  my_dev::dev_mem<uint>  validList(devContext);
-  my_dev::dev_mem<uint>  compactList(devContext);
-
-  int memOffset1 = compactList.cmalloc_copy(localTree.generalBuffer1, localTree.n, 0);
-  int memOffset2 = validList.  cmalloc_copy(localTree.generalBuffer1, localTree.n, memOffset1);
-
-
-  validList.zeroMem();
-  domainCheckSFC.set_arg<int>(0,     &localTree.n);
-  domainCheckSFC.set_arg<uint4>(1,   &lowerBoundary);
-  domainCheckSFC.set_arg<uint4>(2,   &upperBoundary);
-  domainCheckSFC.set_arg<cl_mem>(3,  localTree.bodies_key.p());
-  domainCheckSFC.set_arg<cl_mem>(4,  validList.p());
-  domainCheckSFC.setWork(localTree.n, 128);
-  domainCheckSFC.execute(execStream->s());
-  execStream->sync();
-
-  //Create a list of valid and invalid particles
-  this->resetCompact();  //Make sure compact has been reset
-  int validCount;
-  gpuSplit(devContext, validList, compactList, localTree.n, &validCount);
-
-  LOGF(stderr, "Found %d particles outside my domain, inside: %d \n", validCount, localTree.n-validCount);
-
-//  compactList.d2h();
-//  for(int i=0; i < localTree.n; i++)
-//    LOGF(stderr,"validList: %d \t %d \n", i, compactList[i]);
-
-
-  double tCheck = get_time();
-
-  //Check if the memory size, of the generalBuffer is large enough to store the exported particles
-  //if not allocate more but make sure that the copy of compactList survives
-  int tempSize = localTree.generalBuffer1.get_size() - localTree.n;
-  int needSize = (int)(1.01f*(validCount*(sizeof(bodyStruct)/sizeof(int))));
-  int stepSize = (tempSize / (sizeof(bodyStruct) / sizeof(int)))-512;
-
-  bool doInOneGo = true;
-
-  bodyStruct *extraBodyBuffer = NULL;
-
-  if(stepSize > needSize)
-  {
-    //We can do it in one go
-    doInOneGo = true;
-  }
-  else
-  {
-    //We need an extra CPU buffer
-    doInOneGo       = false;
-    extraBodyBuffer = new bodyStruct[validCount];
-    assert(extraBodyBuffer != NULL);
-  }
-
-  my_dev::dev_mem<bodyStruct>  bodyBuffer(devContext);
-
-  memOffset1 = bodyBuffer.cmalloc_copy(localTree.generalBuffer1, stepSize, memOffset1);
-
-  int extractOffset = 0;
-  for(unsigned int i=0; i < validCount; i+= stepSize)
-  {
-    int items = min(stepSize, (int)(validCount-i));
-
-    if(items > 0)
-    {
-      double t1a = get_time();
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<int>(0,    &extractOffset);
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<int>(1,    &items);
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(2, compactList.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(3, localTree.bodies_Ppos.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(4, localTree.bodies_Pvel.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(5, localTree.bodies_pos.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(6, localTree.bodies_vel.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(7, localTree.bodies_acc0.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(8, localTree.bodies_acc1.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(9, localTree.bodies_time.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(10, localTree.bodies_ids.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(11, localTree.bodies_key.p());
-      extractOutOfDomainParticlesAdvancedSFC.set_arg<cl_mem>(12, bodyBuffer.p());
-      extractOutOfDomainParticlesAdvancedSFC.setWork(items, 128);
-      extractOutOfDomainParticlesAdvancedSFC.execute(execStream->s());
-      execStream->sync();
-      double t2a = get_time();
-
-      LOGF(stderr,"EXTRB: %d \t %lg \n", items, t2a-t1a);
-
-      bodyBuffer.d2h(items); // validCount);
-
-      if(!doInOneGo)
-      {
-        //Do the extra memory copy to the CPU buffer
-        memcpy(&extraBodyBuffer[extractOffset], &bodyBuffer[0], sizeof(bodyStruct)*items);
-        extractOffset += items;
-      }
-    } // if items > 0
-  }//end for
-
-
-
-  if(doInOneGo)
-  {
-    extraBodyBuffer = &bodyBuffer[0]; //Assign correct pointer
-  }
-
-
-  //Now we have to move particles from the back of the array to the invalid spots
-  //this can be done in parallel with exchange operation to hide some time
-
-  //One integer for counting, true-> initialize to zero so counting starts at 0
-  my_dev::dev_mem<uint>  atomicBuff(devContext);
-  memOffset1 = atomicBuff.cmalloc_copy(localTree.generalBuffer1,1, memOffset1);
-  atomicBuff.zeroMem();
-
-  //Internal particle movement
-  internalMoveSFC.set_arg<int>(0,     &validCount);
-  internalMoveSFC.set_arg<int>(1,     &localTree.n);
-  internalMoveSFC.set_arg<uint4>(2,   &lowerBoundary);
-  internalMoveSFC.set_arg<uint4>(3,   &upperBoundary);
-  internalMoveSFC.set_arg<cl_mem>(4,  compactList.p());
-  internalMoveSFC.set_arg<cl_mem>(5,  atomicBuff.p());
-  internalMoveSFC.set_arg<cl_mem>(6,  localTree.bodies_Ppos.p());
-  internalMoveSFC.set_arg<cl_mem>(7,  localTree.bodies_Pvel.p());
-  internalMoveSFC.set_arg<cl_mem>(8,  localTree.bodies_pos.p());
-  internalMoveSFC.set_arg<cl_mem>(9,  localTree.bodies_vel.p());
-  internalMoveSFC.set_arg<cl_mem>(10, localTree.bodies_acc0.p());
-  internalMoveSFC.set_arg<cl_mem>(11, localTree.bodies_acc1.p());
-  internalMoveSFC.set_arg<cl_mem>(12, localTree.bodies_time.p());
-  internalMoveSFC.set_arg<cl_mem>(13, localTree.bodies_ids.p());
-  internalMoveSFC.set_arg<cl_mem>(14, localTree.bodies_key.p());
-  internalMoveSFC.setWork(validCount, 128);
-  internalMoveSFC.execute(execStream->s());
-
-  double tExtract = get_time();
-
-  LOGF(stderr,"Fulll extract OLD took: %lg \n", get_time()-tStart);
-
-  int currentN = localTree.n;
-
-  //this->gpu_exchange_particles_with_overflow_check_SFC(localTree, &bodyBuffer[0], compactList, validCount);
-  this->gpu_exchange_particles_with_overflow_check_SFC(localTree, &extraBodyBuffer[0], compactList, validCount);
-
-
-  double tEnd = get_time();
-
-  char buff5[1024];
-  sprintf(buff5,"EXCHANGE-%d: tCheckDomain: %lg tExtract: %lg tDomainEx: %lg nExport: %d nImport: %d \n",
-      procId, tCheck-tStart, tExtract-tCheck, tEnd-tExtract,validCount, localTree.n - (currentN-validCount));
-  devContext.writeLogEvent(buff5);
-
-  if(!doInOneGo) delete[] extraBodyBuffer;
-
-
-#endif
-
-
-
 
 
 #if 0
@@ -1934,9 +1681,6 @@ void octree::gpuRedistributeParticles_SFC(uint4 *boundaries)
 #endif
 } //End gpuRedistributeParticles
 
-
-#if 1
-
 //Exchange particles with other processes
 int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree,
                                                             bodyStruct *particlesToSend,
@@ -1953,14 +1697,6 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
     recvCount     += nreceive[i];
   }
 
-  #if 0
-    { /* jb2404 */
-      fprintf(stderr, "Proc: %d Exchange, received %d \tSend: %d newN: %d\n",
-          procId, recvCount, nToSend,  tree.n + recvCount - nToSend);
-    }
-  #endif
-
-
   static std::vector<bodyStruct> recv_buffer3;
   recv_buffer3.resize(recvCount);
 
@@ -1969,7 +1705,9 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
   #define NMAXPROC 32768
   static MPI_Status stat[NMAXPROC];
   static MPI_Request req[NMAXPROC*2];
+  assert(nprocs < NMAXPROC);
 
+  //TODO this loop could overflow if scount > INT_MAX (same for rcount)
   int nreq = 0;
   for (int dist = 1; dist < nProcs; dist++)
   {
@@ -1981,7 +1719,6 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
     assert(scount >= 0);
     assert(rcount >= 0);
 
-#if 1
     if (scount > 0)
     {
       MPI_Isend(&particlesToSend[nsendDispls[dst]], scount, MPI_DOUBLE, dst, 1, mpiCommWorld, &req[nreq++]);
@@ -1991,23 +1728,13 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
       MPI_Irecv(&recv_buffer3[recvOffset], rcount, MPI_DOUBLE, src, 1, mpiCommWorld, &req[nreq++]);
       recvOffset += nreceive[src];
     }
-#else
-    MPI_Status stat;
-    MPI_Sendrecv(&particlesToSend[nsendDispls[dst]],
-        scount, MPI_DOUBLE, dst, 1,
-        &recv_buffer3[recvOffset], rcount, MPI_DOUBLE, src, 1, mpiCommWorld, &stat);
-    recvOffset += nreceive[src];
-#endif
   }
 
   double t94 = get_time();
   MPI_Waitall(nreq, req, stat);
-
   double tSendEnd = get_time();
-  //LOGF(stderr, "EXCHANGE V2 comm iter: %d  data-send: %lg data-wait: %lg \n",iter, t94-tStart, tSendEnd -t94);
 
   //If we arrive here all particles have been exchanged, move them to the GPU
-
   LOGF(stderr,"Required inter-process communication time: %lg ,proc: %d\n", get_time()-tStart, procId);
 
   //Compute the new number of particles:
@@ -2015,15 +1742,12 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
 
   LOGF(stderr, "Exchange, received %d \tSend: %d newN: %d\n", recvCount, nToSend, newN);
 
-
-
   //make certain that the particle movement on the device is complete before we resize
   execStream->sync();
 
   double tSyncGPU = get_time();
 
-  //Allocate 10% extra if we have to allocate, to reduce the total number of
-  //memory allocations
+  //Allocate MULTI_GPU_MEM_INCREASE% extra if we have to allocate, to reduce the total number of memory allocations
   int memSize = newN;
   if(tree.bodies_acc0.get_size() < newN)
     memSize = newN * MULTI_GPU_MEM_INCREASE;
@@ -2074,7 +1798,7 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
       //Copy the data from the MPI receive buffers into the GPU-send buffer
 #pragma omp parallel for
         for(int cpIdx=0; cpIdx < items; cpIdx++)
-          bodyBuffer[cpIdx] = recv_buffer3[insertOffset+cpIdx];
+          bodyBuffer[cpIdx] = recv_buffer3[insertOffset+cpIdx]; //TODO can't we just copy directly from recv_buffer3?
 
       bodyBuffer.h2d(items);
 
@@ -2083,12 +1807,12 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
       insertNewParticlesSFC.set_arg<int>(1,    &items);
       insertNewParticlesSFC.set_arg<int>(2,    &tree.n);
       insertNewParticlesSFC.set_arg<int>(3,    &insertOffset);
-      insertNewParticlesSFC.set_arg<cl_mem>(4, localTree.bodies_Ppos.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(5, localTree.bodies_Pvel.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(6, localTree.bodies_pos.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(7, localTree.bodies_vel.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(8, localTree.bodies_acc0.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(9, localTree.bodies_acc1.p());
+      insertNewParticlesSFC.set_arg<cl_mem>(4,  localTree.bodies_Ppos.p());
+      insertNewParticlesSFC.set_arg<cl_mem>(5,  localTree.bodies_Pvel.p());
+      insertNewParticlesSFC.set_arg<cl_mem>(6,  localTree.bodies_pos.p());
+      insertNewParticlesSFC.set_arg<cl_mem>(7,  localTree.bodies_vel.p());
+      insertNewParticlesSFC.set_arg<cl_mem>(8,  localTree.bodies_acc0.p());
+      insertNewParticlesSFC.set_arg<cl_mem>(9,  localTree.bodies_acc1.p());
       insertNewParticlesSFC.set_arg<cl_mem>(10, localTree.bodies_time.p());
       insertNewParticlesSFC.set_arg<cl_mem>(11, localTree.bodies_ids.p());
       insertNewParticlesSFC.set_arg<cl_mem>(12, localTree.bodies_key.p());
@@ -2118,12 +1842,13 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
   return 0;
 }
 
-#endif
 
 
-
-
-//Functions related to the LET Creation and Exchange
+/********************************************************
+ *                                                      *
+ * Functions related to the LET Creation and Exchange   *
+ *                                                      *
+ ********************************************************/
 
 //Broadcast the group-tree structure (used during the LET creation)
 //First we gather the size, so we can create/allocate memory
@@ -2133,10 +1858,8 @@ int octree::gpu_exchange_particles_with_overflow_check_SFC2(tree_structure &tree
 /****** EGABUROV ****/
 void octree::sendCurrentInfoGrpTree()
 {
-  //TODO further clean up by (re)moving never used code
 #ifdef USE_MPI
 
-#if 1  /* new group code, EGABUROV, modified by JBEDORF to get full functional trees *****/
   localTree.boxSizeInfo.waitForCopyEvent();
   localTree.boxCenterInfo.waitForCopyEvent();
 
@@ -2147,15 +1870,15 @@ void octree::sendCurrentInfoGrpTree()
   mpiSync(); //TODO remove
   double tStartGrp = get_time(); //TODO delete
 
-  int nGroupsSmallSet = 0;
-  int nGroupsFullSet  = 0;
+  int nGroupsSmallSet =  0;
+  int nGroupsFullSet  =  0;
   int depthSmallSet   = -1;
   int searchDepthUsed = 99;
 
   //static int tempStartSmall = 0;
   //static int tempStartDepth = 5;
 
-  const int maxNLarge = 512;
+  const int maxNLarge = 512;  //Maximum number of processes to which we send the full group tree.
 
   //static bool modifyNSmall = true;
 
@@ -2163,22 +1886,24 @@ void octree::sendCurrentInfoGrpTree()
                                      //y = startLevelReduction  [0..6],
                                      //z = doTuning             [0: no, 1 tune]
 
+  //TODO This does the tuning of when to send the full group tree and when to send only
+  //the small group. This is based on the previous iteration statistics on what is send
 
   //TODO this version stops tuning after iteration 16. Adjust method to make
   //this a per-process tuning
 
-  std::vector<int2> globalGroupSizeArray(nProcs);     //x is fullGroup, y = smallGroup
+  std::vector<int2> globalGroupSizeArray    (nProcs);  //x is fullGroup, y = smallGroup
   std::vector<int2> globalGroupSizeArrayRecv(nProcs); //x is fullGroup, y = smallGroup
   globalGroupSizeArray[procId] = make_int2(0,0);      //Nothing to ourselves
 
 
-  //Count the number of processes that used the boundary tree
-  //and compute the average level
+  //Count the number of processes that used the boundary tree and compute the average level
   int smallTreeStart = localTree.level_list[localTree.startLevelMin].x;
   int smallTreeEnd   = localTree.level_list[localTree.startLevelMin].y;
 
   if(iter < 8)
   {
+    //For iteration 0 to 8 always send the full tree
     for(int i=0; i < nProcs; i++)
     {
       fullGrpAndLETRequestStatistics[i].x = 0;
@@ -2289,6 +2014,9 @@ void octree::sendCurrentInfoGrpTree()
 //      localTree.level_list[localTree.startLevelMin].x,
 //      localTree.level_list[localTree.startLevelMin].y);
 
+  //Now that we've found the limits of our searches, perform the actual
+  //searches
+
   nGroupsSmallSet =  extractGroupsTreeFullCount2(
                         groupCentre, groupSize,
                         groupMulti, groupBody,
@@ -2321,9 +2049,9 @@ void octree::sendCurrentInfoGrpTree()
   int nGroups = extractGroupsTreeFullCount(
                       &localTree.boxCenterInfo[0],
                       &localTree.boxSizeInfo[0],
-                      localTree.level_list[localTree.startLevelMin].x,
-                      localTree.level_list[localTree.startLevelMin].y,
-                      localTree.n_nodes, 99, depth);
+                       localTree.level_list[localTree.startLevelMin].x,
+                       localTree.level_list[localTree.startLevelMin].y,
+                       localTree.n_nodes, 99, depth);
   nGroupsFullSet = nGroups;
 
   //Assign the fullGroup sizes to process that need more than the smallGroup size
@@ -2433,7 +2161,7 @@ void octree::sendCurrentInfoGrpTree()
   //1) use MPI_Alltoallv
   //2) use combination of allgatherv (for small) + isend/irecv for large
 
-  //Wait for multipoles to copy
+  //Wait for multipole data to be copied
   localTree.multipole.waitForCopyEvent();
 
   static std::vector<real4> fullBoundaryTree;
@@ -2533,9 +2261,8 @@ void octree::sendCurrentInfoGrpTree()
 //         localTree.level_list[localTree.startLevelMin].y);
   }
 
-#if 1
 
-  //MPI_Allgatherv + Isend/IRecv
+  //MPI_Allgatherv for the small tree and Isend/IRecv for the fulltree
   {
     nGroups = SmallBoundaryTree.size();
     MPI_Allgatherv(&SmallBoundaryTree[0], sizeof(real4)*nGroups, MPI_BYTE,
@@ -2545,10 +2272,11 @@ void octree::sendCurrentInfoGrpTree()
 
   double t2 = get_time();
 
-  //Send / recv loop like particle exchange
+  //Send / receive loop like particle exchange
   #define NMAXPROC 32768
   static MPI_Status stat[NMAXPROC];
   static MPI_Request req[NMAXPROC*2];
+  assert(nProcs < NMAXPROC);
 
   int nreq = 0;
   for (int dist = 1; dist < nProcs; dist++)
@@ -2573,50 +2301,6 @@ void octree::sendCurrentInfoGrpTree()
   }
   MPI_Waitall(nreq, req, stat);
 
-#endif
-
-
-
-//  mpiSync();   MPI_Finalize();   exit(0);
-
-#if 0
-  //MPI alltoallv
-
-  std::vector<int> sendOffsets(nProcs);
-  std::vector<int> sendSizes(nProcs);
-
-  static std::vector<real4> bothBoundaryTrees;
-  bothBoundaryTrees.insert(bothBoundaryTrees.end(),SmallBoundaryTree.begin(), SmallBoundaryTree.end());
-  bothBoundaryTrees.insert(bothBoundaryTrees.end(),fullBoundaryTree.begin(),  fullBoundaryTree.end());
-
-  const int fullTreeOffset = bothBoundaryTrees.size()*sizeof(real4);
-  for(int i=0; i < nProcs; i++)
-  {
-    if(globalGroupSizeArray[i].y >= 0)
-    {
-      sendOffsets[i] = 0; //small tree
-      sendSizes[i] = SmallBoundaryTree.size()*sizeof(real4);
-    }
-    else
-    {
-      sendOffsets[i] = fullTreeOffset; //full tree
-      sendSizes[i] = fullBoundaryTree.size()*sizeof(real4);
-    }
-  }
-
-
-
-  MPI_Alltoallv(&bothBoundaryTrees[0],
-                &sendSizes[0],                &sendOffsets[0],  MPI_BYTE,
-                &globalGrpTreeCntSize[0],
-                (int*)&this->globalGrpTreeCount[0], &displacement[0], MPI_BYTE,
-                mpiCommWorld);
-
-
-#endif
-
-
-//
 
   double tEndGrp = get_time();
   char buff5[1024];
@@ -2624,11 +2308,7 @@ void octree::sendCurrentInfoGrpTree()
                  procId, iter, tEndGrp-tStartGrp, nGroupsSmallSet, nGroupsFullSet, nGroupsSmall, nGroupsLarge, t1-t0, t2-t1, tEndGrp-t2, allGatherVSize / (1024*1024.));
   devContext.writeLogEvent(buff5);
 
-
-
 #endif
-#endif
-
 }
 
 
@@ -2648,7 +2328,7 @@ inline int split_node_grav_impbh(
   const _v4si mask = {fullMask, fullMask, fullMask, zeroMask};
   const _v4sf size = __abs(__builtin_ia32_shufps(nodeCOM1, nodeCOM1, 0xFF));
 
-  //mask to prevent NaN signalling / Overflow ? Required to get good pre-SB performance
+  //mask to prevent NaN signaling / Overflow ? Required to get good pre-SB performance
   const _v4sf nodeCOM   = __builtin_ia32_andps(nodeCOM1,   (_v4sf)mask);
   const _v4sf boxCenter = __builtin_ia32_andps(boxCenter1, (_v4sf)mask);
   const _v4sf boxSize   = __builtin_ia32_andps(boxSize1,   (_v4sf)mask);
@@ -4002,122 +3682,7 @@ typedef struct letObject
 
 
 
-int octree::recursiveTopLevelCheck(uint4 checkNode, real4* treeBoxSizes, real4* treeBoxCenters, real4* treeBoxMoments,
-    real4* grpCenter, real4* grpSize, int &DistanceCheck, int &DistanceCheckPP, int maxLevel)
-{
-  int nodeID = checkNode.x;
-  int grpID  = checkNode.y;
-  int endGrp = checkNode.z;
 
-  real4 nodeCOM  = treeBoxMoments[nodeID*3];
-  real4 nodeSize = treeBoxSizes  [nodeID];
-  real4 nodeCntr = treeBoxCenters[nodeID];
-
-  //     LOGF(stderr,"Checking node: %d grpID: %d endGrp: %d depth: %d \n",nodeID, grpID, endGrp, maxLevel);
-
-  int res = maxLevel;
-
-  nodeCOM.w = nodeCntr.w;
-  for(int grp=grpID; grp < endGrp; grp++)
-  {
-    real4 grpcntr = grpCenter[grp];
-    real4 grpsize = grpSize[grp];
-
-    bool split = false;
-    {
-      //         DistanceCheck++;
-      //         DistanceCheckPP++;
-      //Compute the distance between the group and the cell
-      float3 dr = make_float3(fabs((float)grpcntr.x - nodeCOM.x) - (float)grpsize.x,
-          fabs((float)grpcntr.y - nodeCOM.y) - (float)grpsize.y,
-          fabs((float)grpcntr.z - nodeCOM.z) - (float)grpsize.z);
-
-      dr.x += fabs(dr.x); dr.x *= 0.5f;
-      dr.y += fabs(dr.y); dr.y *= 0.5f;
-      dr.z += fabs(dr.z); dr.z *= 0.5f;
-
-      //Distance squared, no need to do sqrt since opening criteria has been squared
-      float ds2    = dr.x*dr.x + dr.y*dr.y + dr.z*dr.z;
-
-      if (ds2     <= fabs(nodeCOM.w))           split = true;
-      if (fabs(ds2 - fabs(nodeCOM.w)) < 10e-04) split = true; //Limited precision can result in round of errors. Use this as extra safe guard
-
-      //         LOGF(stderr,"Node: %d grp: %d  split: %d || %f %f\n", nodeID, grp, split, ds2, nodeCOM.w);
-    }
-    //       LOGF(stderr,"Node: %d grp: %d  split: %d Leaf: %d \t %d %f \n",nodeID, grp, split, nodeCntr.w <= 0,(host_float_as_int(nodeSize.w) == 0xFFFFFFFF), nodeCntr.w);
-
-    if(split)
-    {
-      if(host_float_as_int(nodeSize.w) == 0xFFFFFFFF)
-      {
-        //We want to split, but then we go to deep. So we need a full tree-walk
-        return -1;
-      }
-
-      int child, nchild;
-      int childinfo = host_float_as_int(nodeSize.w);
-      bool leaf = nodeCntr.w <= 0;
-
-      if(!leaf)
-      {
-        //Node
-        child    =    childinfo & 0x0FFFFFFF;           //Index to the first child of the node
-        nchild   = (((childinfo & 0xF0000000) >> 28)) ; //The number of children this node has
-
-        //Process node children
-        for(int y=child; y < child+nchild; y++)
-        {
-          //Go one level deeper into the tree
-          //uint4 nextCheck = make_uint4(y, grp, endGrp, 0);
-          uint4 nextCheck = make_uint4(y, grpID, endGrp, 0);
-          int res2 = recursiveTopLevelCheck(nextCheck, treeBoxSizes, treeBoxCenters, treeBoxMoments,
-              grpCenter, grpSize, DistanceCheck, DistanceCheckPP, maxLevel+1);
-          if(res2 < 0) return -1;
-
-          res = max(res,res2); //Return max level reached
-        }
-      }//!leaf
-      else
-      {
-        //It is a leaf, no need to check any other groups
-        return res;
-      }
-      //No need to check further groups since this one already succeeded
-      grp = endGrp;
-    }//Split
-  }
-  return res;
-}
-
-int octree::recursiveBasedTopLEvelsCheckStart(tree_structure &tree,
-    real4 *treeBuffer,
-    real4 *grpCenter,
-    real4 *grpSize,
-    int startGrp,
-    int endGrp,
-    int &DistanceCheck)
-{
-  //Tree info
-  const int nParticles = host_float_as_int(treeBuffer[0].x);
-  const int nNodes     = host_float_as_int(treeBuffer[0].y);
-
-  //  LOGF(stderr,"Working with %d and %d || %d %d\n", nParticles, nNodes, 1+nParticles+nNodes,nTopLevelTrees );
-
-  real4* treeBoxSizes   = &treeBuffer[1+nParticles];
-  real4* treeBoxCenters = &treeBuffer[1+nParticles+nNodes];
-  real4* treeBoxMoments = &treeBuffer[1+nParticles+2*nNodes];
-
-  const int nodeID = 0;
-  uint4 checkNode = make_uint4(nodeID, startGrp, endGrp, 0);
-
-  int DistanceCheckPP = 0;
-  int maxLevel = recursiveTopLevelCheck(checkNode, treeBoxSizes, treeBoxCenters, treeBoxMoments,
-      grpCenter, grpSize, DistanceCheck, DistanceCheckPP, 0);
-
-
-  //  LOGF(stderr, "Finally Max level found: %d Process : %d \n", maxLevel, ibox)
-  return maxLevel;
-}
 
 
 void octree::checkGPUAndStartLETComputation(tree_structure &tree,
@@ -4131,13 +3696,9 @@ void octree::checkGPUAndStartLETComputation(tree_structure &tree,
                                             int            *treeBuffersSource,
                                             real4         **treeBuffers)
 {
-  //This determines if we interrupt the computation by starting a gravity kernel on the GPU
+  //This determines if we interrupt the LET computation by starting a gravity kernel on the GPU
   if(gravStream->isFinished())
   {
-    double tA1 = get_time(); //TODO DELETE
-//    LOGF(stderr,"GRAVFINISHED %d recvTree: %d  Time: %lg Since start: %lg\n",
-//                 procId, nReceived, get_time()-t1, get_time()-t0);
-
     //Only start if there actually is new data
     if((nReceived - procTrees) > 0)
     {
@@ -4157,19 +3718,16 @@ void octree::checkGPUAndStartLETComputation(tree_structure &tree,
           topNodeCount, recvTree, mergeOwntree, procTrees, tStart);
       LOGF(stderr, "Merging and launchingA iter: %d took: %lg \n", iter, get_time()-t000);
 
-
       //Correct the topNodeOnTheFlyCounter
   #pragma omp critical(updateReceivedProcessed)
       {
         //Compute how many are left, and add these back to the globalCounter
-        int nTopNodesLeft     = oriTopCount-topNodeCount;
-        topNodeOnTheFlyCount += nTopNodesLeft;
+        topNodeOnTheFlyCount += (oriTopCount-topNodeCount);
       }
 
       totalLETExTime += thisPartLETExTime;
-    }// (nReceived - procTrees) > 0)
-  }// isFinished
-
+    }// if (nReceived - procTrees) > 0)
+  }// if isFinished
 }
 
 
@@ -4209,10 +3767,11 @@ void octree::essential_tree_exchangeV2(tree_structure &tree,
   int topNodeOnTheFlyCount = 0;
 
   this->fullGrpAndLETRequestStatistics[procId] = make_int2(0, 0); //Reset our box
+  //For the first 8 iterations mark boundary as used
   for(int i=0; i < nProcs; i++)
   {
     if(iter < 8)
-      this->fullGrpAndLETRequestStatistics[i] = make_int2(1, 1); //Mark as used boundary
+      this->fullGrpAndLETRequestStatistics[i] = make_int2(1, 1);
   }
 
   uint2 node_begend;
@@ -4221,7 +3780,6 @@ void octree::essential_tree_exchangeV2(tree_structure &tree,
 
   int resultOfQuickCheck[nProcs];
 
-  //int2 quickCheckSendSizes [nProcs];
   int4 quickCheckSendSizes [nProcs];
   int  quickCheckSendOffset[nProcs];
 
@@ -5481,10 +5039,6 @@ void octree::ICSend(int destination, real4 *bodyPositions, real4 *bodyVelocities
   MPI_Send( bodyPositions,  toSend*sizeof(real)*4, MPI_BYTE, destination, destination*2+1, mpiCommWorld);
   MPI_Send( bodyVelocities, toSend*sizeof(real)*4, MPI_BYTE, destination, destination*2+2, mpiCommWorld);
   MPI_Send( bodiesIDs,      toSend*sizeof(ullong), MPI_BYTE, destination, destination*2+3, mpiCommWorld);
-
-  /*    MPI_Send( (real*)&bodyPositions[0],  toSend*sizeof(real)*4, MPI_BYTE, destination, destination*2+1, mpiCommWorld);
-        MPI_Send( (real*)&bodyVelocities[0], toSend*sizeof(real)*4, MPI_BYTE, destination, destination*2+2, mpiCommWorld);
-        MPI_Send( (int *)&bodiesIDs[0],      toSend*sizeof(int),    MPI_BYTE, destination, destination*2+3, mpiCommWorld);*/
 #endif
 }
 
@@ -5493,3097 +5047,210 @@ void octree::ICRecv(int recvFrom, vector<real4> &bodyPositions, vector<real4> &b
 #ifdef USE_MPI
   MPI_Status status;
   int nreceive;
-  int procId = mpiGetRank();
 
-  //First send the number of particles, then the actual sample data
+  //First receive the number of particles, then the actual sample data
   MPI_Recv(&nreceive, 1, MPI_INT, recvFrom, procId*2, mpiCommWorld,&status);
 
   bodyPositions.resize(nreceive);
   bodyVelocities.resize(nreceive);
   bodiesIDs.resize(nreceive);
 
-  //Recv the positions, velocities and ids
+  //Receive the positions, velocities and ids
   MPI_Recv( (real*  )&bodyPositions[0],  nreceive*sizeof(real)*4, MPI_BYTE, recvFrom, procId*2+1, mpiCommWorld,&status);
   MPI_Recv( (real*  )&bodyVelocities[0], nreceive*sizeof(real)*4, MPI_BYTE, recvFrom, procId*2+2, mpiCommWorld,&status);
   MPI_Recv( (ullong*)&bodiesIDs[0],      nreceive*sizeof(ullong), MPI_BYTE, recvFrom, procId*2+3, mpiCommWorld,&status);
 #endif
 }
 
+//Sum the number of particles on all processes
 void octree::mpiSumParticleCount(int numberOfParticles)
 {
-  //Sum the number of particles on all processes
+  nTotalFreq_ull = numberOfParticles;
 #ifdef USE_MPI
   unsigned long long tmp  = 0;
   unsigned long long tmp2 = numberOfParticles;
   MPI_Allreduce(&tmp2,&tmp,1, MPI_UNSIGNED_LONG_LONG, MPI_SUM,mpiCommWorld);
   nTotalFreq_ull = tmp;
+#endif
+
+  if(procId == 0) LOG("Total number of particles: %llu\n", nTotalFreq_ull);
+}
+
+
+
+#if 0
+//Uses one communication by storing data in one buffer and communicate required information,
+//such as box-sizes and number of sample particles on this process (Note that the number is only
+//used by process-0
+void octree::sendCurrentRadiusAndSampleInfo(real4 &rmin, real4 &rmax, int nsample, int *nSamples)
+{
+
+  sampleRadInfo curProcState;
+
+  curProcState.nsample      = nsample;
+  curProcState.rmin         = make_double4(rmin.x, rmin.y, rmin.z, rmin.w);
+  curProcState.rmax         = make_double4(rmax.x, rmax.y, rmax.z, rmax.w);
+#ifdef USE_MPI
+  //Get the number of sample particles and the domain size information
+  MPI_Allgather(&curProcState, sizeof(sampleRadInfo), MPI_BYTE,  curSysState,
+      sizeof(sampleRadInfo), MPI_BYTE, mpiCommWorld);
 #else
-  nTotalFreq_ull = numberOfParticles;
+  curSysState[0] = curProcState;
 #endif
 
+  rmin.x                 = (real)(currentRLow[0].x = curSysState[0].rmin.x);
+  rmin.y                 = (real)(currentRLow[0].y = curSysState[0].rmin.y);
+  rmin.z                 = (real)(currentRLow[0].z = curSysState[0].rmin.z);
+  currentRLow[0].w = curSysState[0].rmin.w;
 
-#ifdef PRINT_MPI_DEBUG
-  if(procId == 0)
-    LOG("Total number of particles: %llu\n", nTotalFreq_ull);
-#endif
-}
+  rmax.x                 = (real)(currentRHigh[0].x = curSysState[0].rmax.x);
+  rmax.y                 = (real)(currentRHigh[0].y = curSysState[0].rmax.y);
+  rmax.z                 = (real)(currentRHigh[0].z = curSysState[0].rmax.z);
+  currentRHigh[0].w = curSysState[0].rmax.w;
 
+  nSamples[0] = curSysState[0].nsample;
 
-
-/*
- *
- *
- *
- *  OLD / NON Used code
- *
- *
- *
- *
- */
-
-#if 0
-
-template<class T>
-int octree::MP_exchange_particle_with_overflow_check(int ibox,
-  T *source_buffer,
-  vector<T> &recv_buffer,
-  int firstloc,
-  int nparticles,
-  int isource,
-  int &nsend,
-  unsigned int &recvCount)
-{
-#ifdef USE_MPI
-MPI_Status status;
-int local_proc_id = procId;
-nsend = nparticles;
-
-//first send&get the number of particles to send&get
-unsigned int nreceive;
-
-//Send and get the number of particles that are exchanged
-MPI_Sendrecv(&nsend,1,MPI_INT,ibox,local_proc_id*10,
-    &nreceive,1,MPI_INT,isource,isource*10,mpiCommWorld,
-    &status);
-
-int ss         = sizeof(T);
-int sendoffset = nparticles-nsend;
-
-//Resize the receive buffer
-if((nreceive + recvCount) > recv_buffer.size())
-{
-  recv_buffer.resize(nreceive + recvCount);
-}
-
-
-//Send the actual particles
-MPI_Sendrecv(&source_buffer[firstloc+sendoffset],ss*nsend,MPI_BYTE,ibox,local_proc_id*10+1,
-    &recv_buffer[recvCount],ss*nreceive,MPI_BYTE,isource,isource*10+1,
-    mpiCommWorld,&status);
-
-recvCount += nreceive;
-
-//     int iret = 0;
-//     int giret;
-//     MPI_Allreduce(&iret, &giret,1, MPI_INT, MPI_MAX,mpiCommWorld);
-//     return giret;
-#endif
-return 0;
-}
-
-
-#ifdef USE_MPI
-//SSE optimized MAC check
-inline int split_node_grav_impbh_sse(
-    const _v4sf nodeCOM1,
-    const _v4sf boxCenter1,
-    const _v4sf boxSize1)
-{
-  const _v4si mask = {0xffffffff, 0xffffffff, 0xffffffff, 0x0};
-  const _v4sf size = __abs(__builtin_ia32_shufps(nodeCOM1, nodeCOM1, 0xFF));
-
-  //mask to prevent NaN signalling / Overflow ? Required to get good pre-SB performance
-  const _v4sf nodeCOM   = __builtin_ia32_andps(nodeCOM1,   (_v4sf)mask);
-  const _v4sf boxCenter = __builtin_ia32_andps(boxCenter1, (_v4sf)mask);
-  const _v4sf boxSize   = __builtin_ia32_andps(boxSize1,   (_v4sf)mask);
-
-
-  const _v4sf dr   = __abs(boxCenter - nodeCOM) - boxSize;
-  const _v4sf ds   = dr + __abs(dr);
-  const _v4sf dsq  = ds*ds;
-  const _v4sf t1   = __builtin_ia32_haddps(dsq, dsq);
-  const _v4sf t2   = __builtin_ia32_haddps(t1, t1);
-  const _v4sf ds2  = __builtin_ia32_shufps(t2, t2, 0x00)*(_v4sf){0.25f, 0.25f, 0.25f, 0.25f};
-
-
-  const float c = 10e-4f;
-  const int res = __builtin_ia32_movmskps(
-      __builtin_ia32_orps(
-        __builtin_ia32_cmpleps(ds2,  size),
-        __builtin_ia32_cmpltps(ds2 - size, (_v4sf){c,c,c,c})
-        )
-      );
-
-  //return 1;
-  return res;
-
-}
-
-#endif
-
-//Walk the group tree over the local-data tree.
-//Counts the number of particles and nodes that will be selected
-void octree::tree_walking_tree_stack_versionC13(
-    real4 *multipoleS, nInfoStruct* nodeInfoS, //Local Tree
-    real4* grpNodeSizeInfoS, real4* grpNodeCenterInfoS, //remote Tree
-    int start, int end, int startGrp, int endGrp,
-    int &nAcceptedNodes, int &nParticles,
-    uint2 *curLevel, uint2 *nextLevel)
-{
-#ifdef USE_MPI
-
-  //nodeInfo.z bit values:
-  //  bit 0 : Node has been visit (1)
-  //  bit 1 : Node is split  (2)
-  //  bit 2 : Node is a leaf which particles have been added (4)
-
-  const _v4sf*         multipoleV = (const _v4sf*)        multipoleS;
-  const _v4sf*   grpNodeSizeInfoV = (const _v4sf*)  grpNodeSizeInfoS;
-  const _v4sf* grpNodeCenterInfoV = (const _v4sf*)grpNodeCenterInfoS;
-
-  for(int i=0; i < start; i++) nodeInfoS[i].z = 3;
-
-  nAcceptedNodes = start;
-  nParticles     = 0;
-
-  int curLevelCount  = 0;
-  int nextLevelCount = 0;
-
-  for(int k = start; k < end; k++)
-  {
-    uint2 stackItem;
-    stackItem.x = k;
-    curLevel[curLevelCount++] = stackItem;
-  }
-
-  bool overRuleBegin = true;
-
-  while(curLevelCount > 0)
-  {
-    nextLevelCount = 0;
-    for(int idx = 0; idx < curLevelCount; idx++)
-      //for(int idx = curLevelCount-1; idx >= 0; idx--) //std::stack order
-    {
-      const uint2 stackItem  = curLevel[idx];
-      const uint nodeID      = stackItem.x;
-
-      //Tree-node information
-      const nInfoStruct nodeInfoX = nodeInfoS[nodeID];
-
-      //Early out if this is an accepted leaf node
-      if(nodeInfoX.z  & 4) continue;
-
-      //Only mark the first time, saves writes --> Always writing turns out to be faster
-      if(nodeInfoX.z == 0)
-      {
-        nAcceptedNodes++;
-        nodeInfoS[nodeID].z = 1;
-      }
-
-      //Read the COM and combine it with opening angle criteria from nodeInfoX.x
-      _v4sf nodeCOM = multipoleV[nodeID*3];
-      nodeCOM       = __builtin_ia32_vec_set_v4sf (nodeCOM, nodeInfoX.x, 3);
-
-      int begin, end;
-
-      //I need this since I can't guarantee that I can encode the start-grp info
-      //in the available bytes.
-      if(overRuleBegin)
-      {
-        begin = startGrp; end   = endGrp;
-      }
-      else
-      {
-        begin = stackItem.y & 0x0FFFFFFF;
-        end   = begin +  ((stackItem.y & 0xF0000000) >> 28) ;
-      }
-
-      for(int grpId=begin; grpId <= end; grpId++)
-      {
-        //Group information
-        const _v4sf grpCenter = grpNodeCenterInfoV[grpId];
-        const _v4sf grpSize   = grpNodeSizeInfoV[grpId];
-
-        const int split = split_node_grav_impbh_sse(nodeCOM, grpCenter, grpSize);
-        //        const int split = 1;
-
-        if(split)
-        {
-          const bool leaf        = nodeInfoX.x <= 0;
-
-          if(!leaf)
-          {
-            //nodeInfoS[nodeID].z = nodeInfoS[nodeID].z |  3; //Mark this node as being split, usefull for
-            nodeInfoS[nodeID].z = 3; //Mark this node as being split, useful for
-            //when creating LET tree, we can mark end-points
-            //Sets the split, and visit bits
-
-            const int child    =    nodeInfoX.y & 0x0FFFFFFF;            //Index to the first child of the node
-            const int nchild   = (((nodeInfoX.y & 0xF0000000) >> 28)) ;  //The number of children this node has
-#if 0
-            int childinfoGrp;
-            if( __builtin_ia32_vec_ext_v4sf(grpCenter, 3) <= 0)
-            { //Its a leaf so we stay with this group
-              //              childinfoGrp = grpId | (1) << 28;
-              childinfoGrp = grpId;
-            }
-            else
-              childinfoGrp    = __builtin_ia32_vec_ext_v4si((_v4si)grpSize,3);
-#else
-            int childinfoGrp = grpId;
-            //If its not a leaf so we continue down the group-tree
-            if( __builtin_ia32_vec_ext_v4sf(grpCenter, 3) > 0)
-              childinfoGrp    = __builtin_ia32_vec_ext_v4si((_v4si)grpSize,3);
-#endif
-
-            //Go check the child nodes and child grps
-            for(int i=child; i < child+nchild; i++)
-            { //Add the nodes to the stack
-              nextLevel[nextLevelCount++] = make_uint2(i, childinfoGrp);
-            }
-          }//if !leaf
-          else if(nodeInfoS[nodeID].z != 7)
-          {
-            int nchild  = (((nodeInfoX.y & INVBMASK) >> LEAFBIT)+1);
-
-            //Mark this leaf as completed, no further checks required
-            nodeInfoS[nodeID].z = 7;
-            nParticles        += nchild;
-            break; //We can jump out of these groups
-          }//if !leaf
-        }//if !split
-      }//for grps
-    }// end inner while
-
-    //Swap stacks
-    uint2 *temp = nextLevel;
-    nextLevel   = curLevel;
-    curLevel    = temp;
-
-    curLevelCount  = nextLevelCount;
-    nextLevelCount = 0;
-    overRuleBegin = false;
-  } //end inner while
-#endif //if USE_MPI
-} //end function
-
-
-//Function that walks over the pre-processed data (processed/generated by the tree-tree walk)
-//and fills the LET buffers with the data from the particles and the nodes
-void octree::stackFill(real4 *LETBuffer, real4 *nodeCenter, real4* nodeSize,
-    real4* bodies, real4 *multipole,
-    nInfoStruct *nodeInfo,
-    int nParticles, int nNodes,
-    int start, int end,
-    uint *curLevelStack, uint* nextLevelStack)
-{
-#ifdef USE_MPI
-
-  int curLeveCount   = 0;
-  int nextLevelCount = 0;
-
-  int nStoreIdx = nParticles;
-
-  nParticles = 0;
-
-  int multiStoreIdx = nStoreIdx+2*nNodes; //multipole starts after particles and nodeSize, nodeCenter
-
-  //Copy top nodes, directly after the bodies
-  for(int node=0; node < start; node++)
-  {
-    LETBuffer[nStoreIdx]              = nodeSize[node];
-    LETBuffer[nStoreIdx+nNodes]       = nodeCenter[node];
-    memcpy(&LETBuffer[multiStoreIdx], &multipole[3*node], sizeof(float4)*(3));
-    multiStoreIdx += 3;
-    nStoreIdx++;
-  }
-
-  int childNodeOffset     = end;
-
-  for(int node=start; node < end; node++){
-    curLevelStack[curLeveCount++] = node;
-  }
-
-  while(curLeveCount > 0)
-  {
-    for(int i=0; i < curLeveCount; i++)
-    {
-      const uint node               = curLevelStack[i];
-      const nInfoStruct curNodeInfo = nodeInfo[node];
-      nodeInfo[node].z = 0; //Reset the node for next round/tree
-
-      const int childinfo = curNodeInfo.y;
-      uint newChildInfo   = 0xFFFFFFFF; //Mark node as not split
-
-      uint child, nchild;
-
-      if(curNodeInfo.z & 2) //Split
-      {
-        if(curNodeInfo.z & 4)
-        { //Leaf that is split
-          child   =   childinfo & BODYMASK;
-          nchild  = (((childinfo & INVBMASK) >> LEAFBIT)+1);
-          newChildInfo = nParticles; //Set the start of the particle buffer
-
-          memcpy(&LETBuffer[nParticles], &bodies[child], sizeof(float4)*(nchild));
-          nParticles += nchild;
-
-          nchild  =  nchild-1; //Minus 1 to get it in right format when combining
-          //with newChildInfo
-        }
-        else
-        { //Normal node that is split
-          child    =    childinfo & BODYMASK;           //Index to the first child of the node
-          nchild   = (((childinfo & INVBMASK) >> LEAFBIT)) ; //The number of children this node has
-
-          newChildInfo     = childNodeOffset;
-          childNodeOffset += nchild;
-
-          for(int j=child; j < child+nchild; j++)
-          {
-            nextLevelStack[nextLevelCount++] = j;
-          }//for
-        }//if leaf
-      }//if split
-
-
-      //Copy this node info
-      LETBuffer[nStoreIdx]            = nodeSize[node];
-      LETBuffer[nStoreIdx].w          = host_int_as_float(newChildInfo | (nchild << LEAFBIT));
-      LETBuffer[nStoreIdx+nNodes]     = nodeCenter[node];
-      memcpy(&LETBuffer[multiStoreIdx], &multipole[3*node], sizeof(float4)*(3));
-      multiStoreIdx += 3;
-      nStoreIdx++;
-      if(procId < 0)
-      {
-        if(node < 200)
-          LOGF(stderr, "Node-normal: %d\tMultipole: %f\n", node, multipole[3*node].x);
-      }
-    }//end for
-
-    //Swap stacks
-    uint *temp          = curLevelStack;
-    curLevelStack       = nextLevelStack;
-    nextLevelStack      = temp;
-    curLeveCount        = nextLevelCount;
-    nextLevelCount      = 0;
-  }
-
-#endif //if USE_MPI
-}//stackFill
-
-#endif
-
-#if 0
-
-
-//Exchange the LET structure, this is a point to point communication operation
-real4* octree::MP_exchange_bhlist(int ibox, int isource,
-    int bufferSize, real4 *letDataBuffer)
-{
-#ifdef USE_MPI
-  MPI_Status status;
-  int nrecvlist;
-  int nlist = bufferSize;
-
-  double t0 = get_time();
-  //first send&get the number of particles to send&get
-  MPI_Sendrecv(&nlist,1,MPI_INT,ibox,procId*10, &nrecvlist,
-      1,MPI_INT,isource,isource*10,mpiCommWorld, &status);
-
-  double t1= get_time();
-  //Resize the buffer so it has the correct size and then exchange the tree
-  real4 *recvDataBuffer = new real4[nrecvlist];
-
-  double t2=get_time();
-  //Particles
-  MPI_Sendrecv(&letDataBuffer[0], nlist*sizeof(real4), MPI_BYTE, ibox, procId*10+1,
-      &recvDataBuffer[0], nrecvlist*sizeof(real4), MPI_BYTE, isource, isource*10+1,
-      mpiCommWorld, &status);
-
-  LOG("LET Data Exchange: %d <-> %d  sync-size: %f  alloc: %f  data: %f Total: %lg MB : %f \n",
-      ibox, isource, t1-t0, t2-t1, get_time()-t2, get_time()-t0, (nlist*sizeof(real4)/(double)(1024*1024)));
-
-  return recvDataBuffer;
-#else
-  return NULL;
-#endif
-}
-
-#endif
-
-#if 0
-
-typedef struct hashInfo
-{
-  int     nHashes;      //Number of hashes that will be send by the process
-  int     nParticles;   //Number of particles the sending process has in total
-  double  execTime;     //The time it took this process to compute gravity in the previous step
-  double  execTime2;    //A second timing number. We dont want execTime and execTime2 to fluctuate too much
-  //balance on the one with the largest difference.
-} hashInfo;
-
-int balanceLoad(int *nParticlesOriginal, int *nParticlesNew, float *load,
-    int nProcs, int leftIdx, int nTotal, float loadAvg)
-{
-#ifdef USE_MPI
-  //Sum the total load left and right
-  int nProcsLeftSide    = nProcs / 2;
-  int nProcsRightSide   = nProcs  - nProcsLeftSide;
-  int rightIdx          = leftIdx + nProcsLeftSide;
-
-  if(nProcs == 1) {
-    LOGF(stderr, "Ready by default \n");
-    nParticlesNew[leftIdx] = nTotal;
-    return 0;
-  }
-
-  LOGF(stderr, "Start balance: nProcs: %d, leftIdx: %d, rightIdx: %d nProcLeft: %d  nProcRight: %d nTotal: %d avg: %f\n",
-      nProcs, leftIdx, rightIdx, nProcsLeftSide,nProcsRightSide, nTotal, loadAvg);
-
-  int nPartLeftOriginal = 0, nPartRightOriginal = 0;
-  for(int i=leftIdx;  i < rightIdx;                   i++) nPartLeftOriginal  += nParticlesOriginal[i];
-  for(int i=rightIdx; i < rightIdx+nProcsRightSide;   i++) nPartRightOriginal += nParticlesOriginal[i];
-
-  //Compute the factor to which to increase by using the received timing numbers
-  float loadLeft = 0, loadRight = 0;
-  for(int i=leftIdx;  i < rightIdx;                 i++) loadLeft  += load[i];
-  for(int i=rightIdx; i < rightIdx+nProcsRightSide; i++) loadRight += load[i];
-
-
-  float leftTarget  = (loadAvg*nProcsLeftSide)  / (loadLeft);
-  float rightTarget = (loadAvg*nProcsRightSide) / (loadRight);
-  //Inverse load, for testing LET
-  //LET float leftTarget  = 1./((loadAvg*nProcsLeftSide)  / (loadLeft));
-  //float rightTarget = 1./((loadAvg*nProcsRightSide) / (loadRight));
-
-  int newLeft = 0, newRight = 0;
-  //Check which target we are trying to match, namely the one with minimal work
-  if(leftTarget < rightTarget)
-    //LET  if(leftTarget > rightTarget)
-  {
-    //Optimize left
-    newLeft  = nPartLeftOriginal*leftTarget;
-    newRight = nTotal - newLeft;
-  }
-  else
-  {
-    //Optimize right
-    newRight  = nPartRightOriginal*rightTarget;
-    newLeft   = nTotal - newRight;
-  }
-
-  LOGF(stderr, "newLeft: %d , newRight: %d nTotal: %d , leftTarget: %f rightTarget: %f , loadLeft: %f loadRight: %f \n",
-      newLeft, newRight, nTotal, leftTarget, rightTarget, loadLeft, loadRight);
-
-  if(nProcs == 2)
-  {
-    nParticlesNew[leftIdx] = newLeft;
-    nParticlesNew[rightIdx] = newRight;
-    return 0;
-  }
-
-  //Recursive the left and right parts
-  balanceLoad(nParticlesOriginal, nParticlesNew, load, nProcsLeftSide, leftIdx,  newLeft, loadAvg);
-  balanceLoad(nParticlesOriginal, nParticlesNew, load, nProcsRightSide, rightIdx, newRight, loadAvg);
-
-#endif
-
-  return 0;
-}
-
-
-
-
-void octree::gpu_collect_hashes(int nHashes, uint4 *hashes, uint4 *boundaries, float lastExecTime, float lastExecTime2)
-{
-#ifdef USE_MPI
-  double t0 = get_time();
-
-  hashInfo hInfo;
-  hInfo.nHashes     = nHashes;
-  hInfo.nParticles  = this->localTree.n;
-  hInfo.execTime    = lastExecTime;
-  hInfo.execTime2   = lastExecTime2;
-
-  LOGF(stderr, "Exectime: Proc: %d -> %f \n", procId, hInfo.execTime);
-
-  int       *nReceiveCnts  = NULL;
-  int       *nReceiveDpls  = NULL;
-  float     *execTimes     = NULL;
-  float     *execTimes2    = NULL;
-  hashInfo  *recvHashInfo  = new hashInfo[nProcs];
-
-  //First receive the number of hashes
-  MPI_Gather(&hInfo, sizeof(hashInfo), MPI_BYTE, recvHashInfo, sizeof(hashInfo), MPI_BYTE, 0, mpiCommWorld);
-
-  int    totalNumberOfHashes = 0;
-  float  timeSum, timeSum2   = 0;
-  int    nTotal              = 0;
-  uint4  *allHashes          = NULL;
-
-  //Compute receive offsets (only process 0), total number of particles, total execution time.
-  if(procId == 0)
-  {
-    nReceiveCnts  = new int  [nProcs];
-    nReceiveDpls  = new int  [nProcs];
-    execTimes     = new float[nProcs];
-    execTimes2    = new float[nProcs];
-
-    nReceiveCnts[0]      = recvHashInfo[0].nHashes;
-    execTimes[0]         = recvHashInfo[0].execTime;
-    execTimes2[0]        = recvHashInfo[0].execTime2;
-
-    //Receive counts and displacements
-    totalNumberOfHashes += nReceiveCnts[0];
-    nReceiveCnts[0]      = nReceiveCnts[0]*sizeof(uint4); //Convert to correct data size
-    nReceiveDpls[0]      = 0;
-    nTotal               = recvHashInfo[0].nParticles;
-    timeSum              = recvHashInfo[0].execTime;
-    timeSum2             = recvHashInfo[0].execTime2;
-
-
-    for(int i=1; i < nProcs; i++)
-    {
-      nReceiveCnts[i]      = recvHashInfo[i].nHashes;
-      execTimes[i]         = recvHashInfo[i].execTime;
-      execTimes2[i]        = recvHashInfo[i].execTime2;
-
-      totalNumberOfHashes += nReceiveCnts[i];
-      nReceiveCnts[i]      = nReceiveCnts[i]*sizeof(uint4);
-      nReceiveDpls[i]      = nReceiveDpls[i-1] + nReceiveCnts[i-1];
-
-      nTotal   += recvHashInfo[i].nParticles;
-      timeSum  += recvHashInfo[i].execTime;
-      timeSum2 += recvHashInfo[i].execTime2;
-    }
-    allHashes                      = new uint4[totalNumberOfHashes+1];
-    allHashes[totalNumberOfHashes] = make_uint4(0,0,0,0); //end boundary
-
-
-
-    //Loop so we can decide on which number to balance
-    float avgLoadTime1 = timeSum  / nProcs;
-    float avgLoadTime2 = timeSum2 / nProcs;
-
-    float maxTime1Diff = 0, maxTime2Diff = 0;
-    for(int i=0; i < nProcs; i++)
-    {
-      float temp1 = abs((avgLoadTime1/recvHashInfo[i].execTime)-1);
-      maxTime1Diff = max(temp1, maxTime1Diff);
-      float temp2 = abs((avgLoadTime2/recvHashInfo[i].execTime2)-1);
-      maxTime2Diff = max(temp2, maxTime2Diff);
-    }
-
-    if(0){
-      if(maxTime2Diff > maxTime1Diff)
-      {
-        for(int i=0; i < nProcs; i++)
-        {
-          execTimes[i] = execTimes[2];
-        }
-        timeSum = timeSum2;
-      }}
-
-    LOGF(stderr, "Max diff  Time1: %f\tTime2: %f Proc0: %f \t %f \n",
-        maxTime1Diff, maxTime2Diff, recvHashInfo[0].execTime, recvHashInfo[0].execTime2);
-  } //if procId == 0
-
-  //Collect hashes on process 0
-  MPI_Gatherv(&hashes[0],    nHashes*sizeof(uint4), MPI_BYTE,
-      &allHashes[0], nReceiveCnts,          nReceiveDpls, MPI_BYTE,
-      0, mpiCommWorld);
-
-  //  MPI_Gatherv((procId ? &sampleArray[0] : MPI_IN_PLACE), nsample*sizeof(real4), MPI_BYTE,
-  //              &sampleArray[0], nReceiveCnts, nReceiveDpls, MPI_BYTE,
-  //              0, mpiCommWorld);
-
-  if(procId == 0)
-  {
-    delete[] nReceiveCnts;
-    delete[] nReceiveDpls;
-
-    int       *nPartPerProc  = new int[nProcs];
-
-    //Sort the keys. Use stable_sort (merge sort) since the seperate blocks are already
-    //sorted. This is faster than std::sort (quicksort)
-    //std::sort(allHashes, allHashes+totalNumberOfHashes, cmp_ph_key());
-    std::stable_sort(allHashes, allHashes+totalNumberOfHashes, cmp_ph_key());
-
-
-#define LOAD_BALANCE 0
-#define LOAD_BALANCE_MEMORY 0
-
-#if LOAD_BALANCE
-    //Load balancing version, based on gravity approximation execution times.
-
-    LOGF(stderr, "Time sum: %f \n", timeSum);
-
-    //Normalize, fractions : timeSum / gravTime -> gives a relative time number
-    float normSum = 0;
-    for(int i=0; i < nProcs; i++){
-      // execTimes[i] = timeSum / execTimes[i];
-      // normSum     += execTimes[i];
-      LOGF(stderr, "Exec after norm: %d\t %f \tn: %d \n",i, execTimes[i],recvHashInfo[i].nParticles );
-    }
-    LOGF(stderr, "Normalized sum:%f  \n",normSum);
-
-
-    int *npartPerProcOld = new int[nProcs];
-    float *loadPerProc   = new float[nProcs];
-
-    for(int i=0; i < nProcs; i++)
-    {
-      npartPerProcOld[i] = recvHashInfo[i].nParticles;
-      loadPerProc[i] = recvHashInfo[i].execTime;
-    }
-    float loadAvg = timeSum / nProcs;
-
-    //Adjust the boundary locations
-    balanceLoad(npartPerProcOld,nPartPerProc, loadPerProc,
-        nProcs,0,nTotal,loadAvg);
-
-    delete[] npartPerProcOld;
-    delete[] loadPerProc;
-    //End adjusting
-
-
-
-    //Compute the number of particles to be assign per process
-    for(int i=0; i < nProcs; i++){
-      //        nPartPerProc[i] = (execTimes[i] / normSum) * nTotal;
-
-      //  nPartPerProc[i] = recvHashInfo[i].nParticles*(2*(execTimes[i] / normSum));
-      //fprintf(stderr, "Npart per proc: %d\t %d \n",i, nPartPerProc[i]);
-    }
-
-    //Average with number of particles of previous step
-    //TODO
-    //float fac = 0.25; //25% old, 75% new
-    float fac = 0.50; //Average
-    for(int i=0; i < nProcs; i++){
-      LOGF(stderr, "Npart per proc: new %d\told %d (avg final: %d)\n",
-          nPartPerProc[i], recvHashInfo[i].nParticles,
-          ((int)((recvHashInfo[i].nParticles*fac) + (nPartPerProc[i] *(1-fac)))));
-      nPartPerProc[i] = (int)((recvHashInfo[i].nParticles*fac) + (nPartPerProc[i] *(1-fac)));
-
-    }
-
-
-    delete[] execTimes;
-    delete[] execTimes2;
-    //Now try to adjust this with respect to memory load-balance
-
-    bool doPrint          = true;
-    bool doMemLoadBalance = (LOAD_BALANCE_MEMORY) ? true : false;
-    if(doMemLoadBalance)
-    {
-      const int maxTries          = std::max(10,nProcs);
-      const double maxDiff        = 1.9; //Each process has to be within a factor 1.9 of the maximum load
-      const int allowedDifference = 5;
-
-      for(int tries = maxTries; tries > 0; tries--)
-      {
-        int maxNumber       = 0;
-        int minNumber       = nPartPerProc[0];
-        int countOutsideMin = 0;
-        int sum             = 0;
-
-        if(doPrint){
-          LOGF(stderr,"Before memory load adjustment: \n");
-          for(int i=0; i < nProcs; i++){
-            LOGF(stderr, "%d \t %d \n", i, nPartPerProc[i]);
-            sum += nPartPerProc[i];
-          }
-          LOGF(stderr, "Sum: %d \n", sum);
-        }
-
-        //First find the max and min process load and compute the minimum number of
-        //particles a process should have according to the defined balance
-        maxNumber = countOutsideMin = 0;
-
-        for(int i=0; i < nProcs; i++){
-          maxNumber = std::max(maxNumber, nPartPerProc[i]);
-          minNumber = std::min(minNumber, nPartPerProc[i]);
-        }
-
-        double requiredNumber = maxNumber / maxDiff;
-
-        if(doPrint){
-          LOGF(stderr, "Max: %d  Min: %d , maxDiff factor: %f  required: %f \n",
-              maxNumber, minNumber, maxDiff, requiredNumber);
-        }
-
-        if((abs(minNumber-requiredNumber)) <= allowedDifference)
-        {
-          break; //Difference between required and achieved is within what we allow, accept this
-        }
-
-        //Count the number of procs below the minNumber and compute number within the limits
-        for(int i=0; i < nProcs; i++) if(requiredNumber > nPartPerProc[i]) countOutsideMin++;
-        int countInsideMin = nProcs - countOutsideMin;
-
-        if(countOutsideMin == 0)
-          break; //Success, all within the range minimum required...maximum assigned
-
-        //Compute particles to be added  to the processes outside the range, we take
-        //(0.5*(Particles Required - minLoad)) / (# processes outside the min range) .
-        //To evenly add particles , to prevent large jumps when # outside is small we use
-        //the factor 0.5. For the particles to be removed we do the same but now use
-        //number of processes inside the min range.
-        int addNumberOfParticles = (0.5*(requiredNumber - minNumber)) / countOutsideMin;
-        addNumberOfParticles     = std::max(1, addNumberOfParticles);
-
-        int removeNumberOfParticles = (0.5*(requiredNumber - minNumber)) / countInsideMin;
-        removeNumberOfParticles     = std::max(1, removeNumberOfParticles);
-
-        if(doPrint){
-          LOGF(stderr, "#Outside: %d , #Inside: %d Adding total: %f per proc: %d  Removing total: %f per proc: %d \n",
-              countOutsideMin, countInsideMin, requiredNumber-minNumber,
-              addNumberOfParticles, requiredNumber - minNumber, removeNumberOfParticles);
-        }
-
-        //Finally modify the particle counts :-)
-        for(int i=0; i < nProcs; i++)
-        {
-          if(nPartPerProc[i] < requiredNumber)
-            nPartPerProc[i] += addNumberOfParticles;
-          else if(nPartPerProc[i] > requiredNumber)
-            nPartPerProc[i] -= removeNumberOfParticles;
-        }//end modify
-
-        if(doPrint){
-          LOGF(stderr,"After memory load adjustment: \n");
-          for(int i=0; i < nProcs; i++) fprintf(stderr, "%d \t %d \n", i, nPartPerProc[i]);
-          LOGF(stderr, "Tries left: %d \n\n\n", tries);
-        }
-      }//for tries
-    } //if doMemLoadBalance
-#else //#if LOAD_BALANCE
-    //Per process, equal number of particles
-    int nPerProc = nTotal / nProcs;
-    for(int i=0; i < nProcs; i++) nPartPerProc[i] = nPerProc;
-    LOGF(stderr, "Number of particles per process: %d \t %d \n", nTotal, nPerProc);
-#endif
-
-
-    //All set and done, get the boundaries
-    int tempSum   = 0;
-    int procIdx   = 1;
-    boundaries[0] = make_uint4(0x0, 0x0, 0x0, 0x0);
-    for(int i=0; i < totalNumberOfHashes; i++)
-    {
-      tempSum += allHashes[i].w;
-      if(tempSum >= nPartPerProc[procIdx-1])
-      {
-        LOGF(stderr, "Boundary at: %d\t%d %d %d %d \t %d \n",
-            i, allHashes[i+1].x,allHashes[i+1].y,allHashes[i+1].z,allHashes[i+1].w, tempSum);
-        tempSum = 0;
-        boundaries[procIdx++] = allHashes[i+1];
-      }
-    }//for totalNumberOfHashes
-
-
-    //Force final boundary to be the highest possible key value
-    boundaries[nProcs]  = make_uint4(0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF, 0xFFFFFFFF);
-
-    delete[] nPartPerProc;
-    delete[] allHashes;
-  }//if procId == 0
-  delete[] recvHashInfo;
-
-
-  //Send the boundaries to all processes
-  MPI_Bcast(boundaries,  sizeof(uint4)*(nProcs+1),MPI_BYTE,0,mpiCommWorld);
-
-  if(procId == 0){
-    for(int i=0; i < nProcs; i++)
-    {
-      LOGF(stderr, "Proc: %d Going from: >= %u %u %u  to < %u %u %u \n",i,
-          boundaries[i].x,boundaries[i].y,boundaries[i].z,boundaries[i+1].x,boundaries[i+1].y,boundaries[i+1].z);
-    }
-    LOGF(stderr, "Exchanging and sorting of hashes took: %f \n", get_time()-t0);
-  }
-
-
-
-
-#endif //if USE_MPI
-  //exit(0);
-}
-
-// === Old essential tree functions ====
-
-#if 0
-#ifdef USE_GROUP_TREE
-  //Build a tree-structure on top of the received data. This to
-  //perform a faster quickCheck :-)
-double tBuildStart = get_time();
-  //Use the root-node only to build a global tree-structure
-  static std::vector<real4> rootCntrs(nProcs-1);
-  static std::vector<real4> rootSizes(nProcs-1);
-  static std::vector< int>  domainIds(nProcs-1);
-
-  for(int ib=0; ib < nProcs-1; ib++)
-  {
-    int ibox  = (nProcs-1)-(ib%nProcs);
-    ibox      = (ibox+procId)%nProcs; //index to send...
-    //Group info for this process
-    int idx         = globalGrpTreeOffsets[ibox];
-    rootCntrs[ib]   = globalGrpTreeCntSize[idx];
-    idx            += this->globalGrpTreeCount[ibox] / 2; //Divide by two to get halfway
-    rootSizes[ib]   = globalGrpTreeCntSize[idx];
-
-    domainIds[ib] = ibox;
-  }
-
-
-  std::vector<real4> treeProperties;
-  std::vector<int  > reorderIDs(domainIds);
-
-
-  double t200 = get_time();
-  const HostConstruction hostConstruct(rootCntrs, rootSizes, treeProperties, reorderIDs, tree.corner);
-  double t210 = get_time();
-  LOGF(stderr,"Building tree-structure took: %lg  Number of nodes/leafs/particles: %d Original roots: %d\n",
-          t210-t200, (int)treeProperties.size()/2, (int)rootCntrs.size());
-
-
-  //end tree-building section
-
-  //Perform a quick test, without keeping statistics
-  //to test how long it takes to traverse the structure
-  {
-    const int tid = 0;
-    const int allocSize = (int)(tree.n_nodes*1.10);
-    getLETBuffers[tid].LETBuffer_node.reserve(allocSize);
-    getLETBuffers[tid].LETBuffer_ptcl.reserve(allocSize);
-    getLETBuffers[tid].currLevelVecI.reserve(allocSize);
-    getLETBuffers[tid].nextLevelVecI.reserve(allocSize);
-    getLETBuffers[tid].currLevelVecUI4.reserve(allocSize);
-    getLETBuffers[tid].nextLevelVecUI4.reserve(allocSize);
-    getLETBuffers[tid].currGroupLevelVec.reserve(allocSize);
-    getLETBuffers[tid].nextGroupLevelVec.reserve(allocSize);
-    getLETBuffers[tid].groupSplitFlag.reserve(allocSize);
-    const int nnodes = treeProperties.size() / 2;
-    unsigned long long nflops;
-    double bla = get_time();
-    double bla3;
-
-    const int sizeTree=  getLEToptQuickCombinedTree(
-                                    quickCheckData[procId],
-                                    getLETBuffers[tid],
-                                    NCELLMAX,
-                                    NDEPTHMAX,
-                                    &nodeCenterInfo[0],
-                                    &nodeSizeInfo[0],
-                                    &multipole[0],
-                                    0, //Cellbeg
-                                    1,  //Cell end
-                                    &bodies[0],
-                                    tree.n,
-                                    &treeProperties[nnodes],    //size
-                                    &treeProperties[0],  //centre
-                                    0,//grp begin
-                                    1,//grp eend
-                                    tree.n_nodes,
-                                    procId, procId,
-                                    nflops, bla3);
-
-    char buff5[1024];
-    sprintf(buff5,"CLETTIME-%d: tTreeTree: %lg\n", procId, get_time()-bla);
-    devContext.writeLogEvent(buff5); //TODO DELETE
-    sprintf(buff5,"DLETTIME-%d: tBuildTree: %lg\n", procId, bla-tBuildStart);
-    devContext.writeLogEvent(buff5); //TODO DELETE
-    sprintf(buff5,"ELETTIME-%d: tTreeTreeSize: %d\n", procId, sizeTree);
-    devContext.writeLogEvent(buff5); //TODO DELETE
-
-    quickCheckData[procId].clear();
-}
-mpiSync();
-  //end tree-building section
-#endif //ifdef USE_GROUP_TREE
-
-
-void extractGroupsTree(
-    std::vector<real4> &groupCentre,
-    std::vector<real4> &groupSize,
-    std::vector<int> grpIdsNormal,
-    const real4 *nodeCentre,
-    const real4 *nodeSize,
-    const int cellBeg,
-    const int cellEnd,
-    const int nNodes)
-{
-  groupCentre.clear();
-  groupCentre.reserve(nNodes);
-
-  groupSize.clear();
-  groupSize.reserve(nNodes);
-
-  static std::vector<int> grpIds;
-  grpIds.clear(); grpIds.reserve(nNodes);
-
-
-
-  const int levelCountMax = nNodes;
-  std::vector<int> currLevelVec, nextLevelVec;
-  currLevelVec.reserve(levelCountMax);
-  nextLevelVec.reserve(levelCountMax);
-  Swap<std::vector<int> > levelList(currLevelVec, nextLevelVec);
-
-  //These are top level nodes. And everything before
-  //should be added. Nothing has to be changed
-  //since we keep the structure
-  for(int cell = 0; cell < cellBeg; cell++)
-  {
-    groupCentre.push_back(nodeCentre[cell]);
-    groupSize.push_back  (nodeSize[cell]);
-    grpIds.push_back(cell);
-  }
-
-  for (int cell = cellBeg; cell < cellEnd; cell++)
-    levelList.first().push_back(cell);
-
-
-  int childOffset = cellEnd;
-
-  int depth = 0;
-  while (!levelList.first().empty())
-  {
-//    LOGF(stderr, " depth= %d Store offset: %d cursize: %d\n", depth++, childOffset, groupSize.size());
-    const int csize = levelList.first().size();
-    for (int i = 0; i < csize; i++)
-    {
-      const uint   nodeIdx = levelList.first()[i];
-      const float4 centre  = nodeCentre[nodeIdx];
-      const float4 size    = nodeSize[nodeIdx];
-      const float nodeInfo_x = centre.w;
-      const uint  nodeInfo_y = host_float_as_int(size.w);
-
-//      LOGF(stderr,"BeforeWorking on %d \tLeaf: %d \t %f [%f %f %f]\n",nodeIdx, nodeInfo_x <= 0.0f, nodeInfo_x, centre.x, centre.y, centre.z);
-
-      const bool lleaf = nodeInfo_x <= 0.0f;
-      if (!lleaf)
-      {
-        const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-        const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-#if 1
-        //We mark this as an end-point
-        if (lnchild == 8)
-        {
-          float4 centre1 = centre;
-          centre1.w = -1; //TODO this has to be some clearer identified value like 0xFFFFF etc
-          groupCentre.push_back(centre1);
-          groupSize  .push_back(size);
-          grpIds.push_back(nodeIdx);
-        }
-        else
-#endif
-        {
-//          LOGF(stderr,"ORIChild info: Node: %d stored at: %d  info:  %d %d \n",nodeIdx, groupSize.size(), lchild, lnchild);
-          //We pursue this branch, mark the offsets and add the parent
-          //to our list and the children to next level process
-          float4 size1   = size;
-          uint newOffset   = childOffset | ((uint)(lnchild) << LEAFBIT);
-          childOffset     += lnchild;
-          size1.w         = host_int_as_float(newOffset);
-
-          groupCentre.push_back(centre);
-          groupSize  .push_back(size1);
-          grpIds.push_back(nodeIdx);
-
-          for (int i = lchild; i < lchild + lnchild; i++)
-            levelList.second().push_back(i);
-        }
-      }
-      else
-      {
-        float4 centre1 = centre;
-        centre1.w = -1;
-        groupCentre.push_back(centre1);
-        groupSize  .push_back(size);
-        grpIds.push_back(nodeIdx);
-      }
-    }
-
-//    LOGF(stderr, "  done depth= %d Store offset: %d cursize: %d\n", depth, childOffset, groupSize.size());
-    levelList.swap();
-    levelList.second().clear();
-  }
-
-
-#if 0 //Verification during testing, compare old and new method
-
-//
-//  char buff[20*128];
-//  sprintf(buff,"Proc: ");
-//  for(int i=0; i < grpIds.size(); i++)
-//  {
-//    sprintf(buff,"%s %d, ", buff, grpIds[i]);
-//  }
-//  LOGF(stderr,"%s \n", buff);
-
-
-  //Verify our results
-  int checkCount = 0;
-  for(int j=0; j < grpIdsNormal.size(); j++)
-  {
-    for(int i=0; i < grpIds.size(); i++)
-    {
-        if(grpIds[i] == grpIdsNormal[j])
-        {
-          checkCount++;
-          break;
-        }
-    }
-  }
-
-  if(checkCount == grpIdsNormal.size()){
-    LOGF(stderr,"PASSED grpTest %d \n", checkCount);
-  }else{
-    LOGF(stderr, "FAILED grpTest %d \n", checkCount);
-  }
-
-
-  std::vector<real4> groupCentre2;
-  std::vector<real4> groupSize2;
-  std::vector<int> grpIdsNormal2;
-
-  extractGroupsPrint(
-     groupCentre2,
-     groupSize2,
-     grpIdsNormal2,
-     &groupCentre[0],
-     &groupSize[0],
-     cellBeg,
-     cellEnd,
-     nNodes);
-
-#endif
-
-}
-
-
-void extractGroupsPrint(
-    std::vector<real4> &groupCentre,
-    std::vector<real4> &groupSize,
-    std::vector<int> &grpIds,
-    const real4 *nodeCentre,
-    const real4 *nodeSize,
-    const int cellBeg,
-    const int cellEnd,
-    const int nNodes)
-{
-  groupCentre.clear();
-  groupCentre.reserve(nNodes);
-
-  groupSize.clear();
-  groupSize.reserve(nNodes);
-
-
-  grpIds.clear(); grpIds.reserve(nNodes);
-
-
-  const int levelCountMax = nNodes;
-  std::vector<int> currLevelVec, nextLevelVec;
-  currLevelVec.reserve(levelCountMax);
-  nextLevelVec.reserve(levelCountMax);
-  Swap<std::vector<int> > levelList(currLevelVec, nextLevelVec);
-
-  for (int cell = cellBeg; cell < cellEnd; cell++)
-    levelList.first().push_back(cell);
-
-//  LOGF(stderr,"Added all start cells: %d to %d \n", cellBeg, cellEnd);
-  int depth = 0;
-  while (!levelList.first().empty())
-  {
-//    LOGF(stderr, " depth= %d \n", depth++);
-    const int csize = levelList.first().size();
-    for (int i = 0; i < csize; i++)
-    {
-      const uint   nodeIdx = levelList.first()[i];
-
-      const float4 centre  = nodeCentre[nodeIdx];
-      const float4 size    = nodeSize[nodeIdx];
-      const float nodeInfo_x = centre.w;
-      const uint  nodeInfo_y = host_float_as_int(size.w);
-
-//      LOGF(stderr,"Working on %d \tLeaf: %d \t %f [%f %f %f]\n",nodeIdx, nodeInfo_x <= 0.0f, nodeInfo_x, centre.x, centre.y, centre.z);
-
-      const bool lleaf = nodeInfo_x <= 0.0f;
-      if (!lleaf)
-      {
-        const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-        const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-#if 1
-//        LOGF(stderr,"Child info: %d %d \n", lchild, lnchild);
-        if (lnchild == 8)
-        {
-          float4 centre1 = centre;
-          centre1.w = -1;
-          groupCentre.push_back(centre1);
-          groupSize  .push_back(size);
-          grpIds.push_back(nodeIdx);
-        }
-        else
-#endif
-          for (int i = lchild; i < lchild + lnchild; i++)
-            levelList.second().push_back(i);
-      }
-      else
-      {
-        float4 centre1 = centre;
-        centre1.w = -1;
-        groupCentre.push_back(centre1);
-        groupSize  .push_back(size);
-        grpIds.push_back(nodeIdx);
-      }
-    }
-
-    levelList.swap();
-    levelList.second().clear();
-  }
-}
-
-
-#if 0 //Old code
-#if 1
-    double tGrpTree = get_time();
-    static std::vector<real4> groupCentre2, groupSize2;
-
-    extractGroupsTree(
-        groupCentre2, groupSize2, grpIds,
-        &localTree.boxCenterInfo[0],
-        &localTree.boxSizeInfo[0],
-        localTree.level_list[localTree.startLevelMin].x,
-        localTree.level_list[localTree.startLevelMin].y,
-        localTree.n_nodes);
-
-    int nGroups2 = groupCentre2.size();
-    LOGF(stderr, "First: %d %d ExtractGroupsTree n: %d [%d] \tTook: %lg \n",
-        localTree.level_list[localTree.startLevelMin].x,
-        localTree.level_list[localTree.startLevelMin].y,
-        nGroups2, (int)groupSize2.size(),
-        get_time() - tGrpTree);
-
-    groupCentre2.insert(groupCentre2.end(), groupSize2.begin(), groupSize2.end());
-    nGroups = groupCentre2.size();
-    groupCentre = groupCentre2;
-#endif
-
-
-  std::vector<int> globalSizeArray(nProcs), displacement(nProcs,0);
-  MPI_Allgather(&nGroups,  sizeof(int), MPI_BYTE,
-      &globalSizeArray[0], sizeof(int), MPI_BYTE, mpiCommWorld); /* to globalSize Array */
-
-  int runningOffset = 0;
-  for (int i = 0; i < nProcs; i++)
-  {
-    this->globalGrpTreeCount[i] = globalSizeArray[i];
-    this->globalGrpTreeOffsets[i] = runningOffset;
-    fullGrpAndLETRequest[i] = 0;
-
-    displacement[i] = runningOffset*sizeof(real4);
-
-    runningOffset += globalSizeArray[i];
-    globalSizeArray[i] *= sizeof(real4);
-  }
-
-
-
-  if (globalGrpTreeCntSize) delete[] globalGrpTreeCntSize;
-  const int totalNumberOfGroups = runningOffset;  /*check if defined */
-  globalGrpTreeCntSize = new real4[totalNumberOfGroups]; /* totalNumberOfGroups = 2*nGroups_recvd */
-
-  /* compute displacements for allgatherv */
-  MPI_Allgatherv(
-      &groupCentre[0], sizeof(real4)*nGroups, MPI_BYTE,
-      globalGrpTreeCntSize, &globalSizeArray[0], &displacement[0], MPI_BYTE,
-      mpiCommWorld);
-
-
-  double tEndGrp = get_time(); //TODO delete
-  char buff5[1024];
-  sprintf(buff5,"BLETTIME-%d: tGrpSend: %lg\n", procId, tEndGrp-tStartGrp);
-  devContext.writeLogEvent(buff5); //TODO DELETE
-  sprintf(buff5,"GLETTIME-%d: nGrpSize: %d\n", procId, nGroups);
-    devContext.writeLogEvent(buff5); //TODO DELETE
-
-
-
-#if 1
-    mpiSync();//TODO delete
-    double tGrpTreeFull = get_time();
-    static std::vector<real4> groupCentre3, groupSize3;
-    static std::vector<real4> groupMulti, groupBody;
-
-    localTree.multipole.waitForCopyEvent();
-
-    extractGroupsTreeFull(
-        groupCentre3, groupSize3,
-        groupMulti, groupBody,
-        grpIds,
-        &localTree.boxCenterInfo[0],
-        &localTree.boxSizeInfo[0],
-        &localTree.multipole[0],
-        &localTree.bodies_Ppos[0],
-        localTree.level_list[localTree.startLevelMin].x,
-        localTree.level_list[localTree.startLevelMin].y,
-        localTree.n_nodes);
-
-    int nGroups3 = groupCentre3.size();
-
-    LOGF(stderr, "ExtractGroupsTreeFull n: %d [%d] Multi: %d \tTook: %lg \n",
-           nGroups3, (int)groupSize3.size(), (int)groupMulti.size(),
-           get_time() - tGrpTreeFull);
-    assert(nGroups3*3 == groupMulti.size());
-
-    //Merge all data into a single array, store offsets
-    const int nbody = groupBody.size();
-    const int nnode = groupSize3.size();
-
-    const int combinedSize = 1 + nbody + 5*nnode;
-    static std::vector<real4> fullBoundaryTree;
-
-    fullBoundaryTree.clear();
-    fullBoundaryTree.reserve(combinedSize);
-    float4 description;
-
-    //Set the tree properties, before we exchange the data
-    description.x = host_int_as_float(nbody);
-    description.y = host_int_as_float(nnode);
-    description.z = host_int_as_float(localTree.level_list[localTree.startLevelMin].x);
-    description.w = host_int_as_float(localTree.level_list[localTree.startLevelMin].y);
-
-    fullBoundaryTree.push_back(description);
-    fullBoundaryTree.insert(fullBoundaryTree.end(), groupBody.begin()   , groupBody.end());
-    fullBoundaryTree.insert(fullBoundaryTree.end(), groupCentre3.begin(), groupCentre3.end());
-    fullBoundaryTree.insert(fullBoundaryTree.end(), groupSize3.begin()  , groupSize3.end());
-    fullBoundaryTree.insert(fullBoundaryTree.end(), groupMulti.begin()  , groupMulti.end());
-
-    assert(fullBoundaryTree.size() == combinedSize);
-
-    int nGroupsFull = fullBoundaryTree.size();
-
-
-    //Exchange these full-trees
-    std::vector<int> globalSizeArrayFull(nProcs), displacementFull(nProcs,0);
-    std::vector<int> globalGrpTreeCountFull(nProcs), globalGrpTreeOffsetsFull(nProcs);
-
-
-    double tGrpFullMid = get_time();
-    mpiSync();
-    double tGrpFullMid2 = get_time();
-
-     MPI_Allgather(&nGroupsFull,            sizeof(int), MPI_BYTE,
-                   (void*)&globalSizeArrayFull[0], sizeof(int), MPI_BYTE,
-                   mpiCommWorld); /* to globalSize Array */
-
-     int runningOffsetFull = 0;
-     for (int i = 0; i < nProcs; i++)
-     {
-       globalGrpTreeCountFull[i]   = globalSizeArrayFull[i];
-       globalGrpTreeOffsetsFull[i] = runningOffsetFull;
-
-       displacementFull[i] = runningOffsetFull*sizeof(real4);
-
-       runningOffsetFull += globalSizeArrayFull[i];
-       globalSizeArrayFull[i] *= sizeof(real4);
-     }
-
-//     LOGF(stderr,"Total number of groups being send around with full trees: %d \n", runningOffsetFull)
-
-
-     const int totalNumberOfGroupsFull = runningOffsetFull;  /*check if defined */
-     if(procId == 0) fprintf(stderr,"Proc: %d Total number of groups full: %d  Time so far: %lg\n", procId,runningOffsetFull, get_time()-tGrpFullMid2);
-//     static std::vector<real4> globalGrpTreeCntSizeFull;
-//     globalGrpTreeCntSizeFull.resize(totalNumberOfGroupsFull); /* totalNumberOfGroups = 2*nGroups_recvd */
-
-     real4 *globalGrpTreeCntSizeFull = new real4[totalNumberOfGroupsFull]; /* totalNumberOfGroups = 2*nGroups_recvd */
-
-
-     /* compute displacements for allgatherv */
-     MPI_Allgatherv(
-         &fullBoundaryTree[0], sizeof(real4)*nGroupsFull, MPI_BYTE,
-         &globalGrpTreeCntSizeFull[0], &globalSizeArrayFull[0], &displacementFull[0], MPI_BYTE,
-         mpiCommWorld);
-
-
-     double tEndGrpFull = get_time(); //TODO delete
-
-     sprintf(buff5,"HLETTIME-%d: tGrpFullPrep: %lg\n", procId, tGrpFullMid-tGrpTreeFull);
-     devContext.writeLogEvent(buff5); //TODO DELETE
-     sprintf(buff5,"HLETTIME-%d: tGrpSendFull: %lg\n", procId, tEndGrpFull-tGrpFullMid2);
-     devContext.writeLogEvent(buff5); //TODO DELETE
-     sprintf(buff5,"ILETTIME-%d: nGrpSizeFull: %d\n", procId, nGroupsFull);
-     devContext.writeLogEvent(buff5); //TODO DELETE
-
-     if(1)
-     {
-
-       static GETLETBUFFERS bufferStructTemp[16];
-
-
-       omp_set_num_threads(16);
-       int omp_ticket = 0;
-       int countGroupsSuccess = 0;
-       double tStartNBoundaryOk = get_time();
-#pragma omp parallel
-       {
-         int tid = omp_get_thread_num();
-         //Test our local boundaries against the received tree
-         while(1)
-         {
-           int ibox;
-           #pragma omp critical
-             ibox = omp_ticket++; //Get a unique ticket
-
-           if(ibox >= nProcs) break;
-
-           if(ibox == procId) continue; //Skip ourself :)
-
-
-           const int allocSize = (int)(16384);
-           bufferStructTemp[tid].LETBuffer_node.reserve(allocSize);
-           bufferStructTemp[tid].LETBuffer_ptcl.reserve(allocSize);
-           bufferStructTemp[tid].currLevelVecI.reserve(allocSize);
-           bufferStructTemp[tid].nextLevelVecI.reserve(allocSize);
-           bufferStructTemp[tid].currLevelVecUI4.reserve(allocSize);
-           bufferStructTemp[tid].nextLevelVecUI4.reserve(allocSize);
-           bufferStructTemp[tid].currGroupLevelVec.reserve(allocSize);
-           bufferStructTemp[tid].nextGroupLevelVec.reserve(allocSize);
-           bufferStructTemp[tid].groupSplitFlag.reserve(allocSize);
-
-
-           int nGroupsLocal            = nGroups;
-           const real4 *groupLocalCntr = &groupCentre2[0];
-           const real4 *groupLocalSize = &groupCentre2[nGroupsLocal/2];
-
-           //Read properties of remote struct
-
-           const int remoteNCount = globalGrpTreeCountFull[ibox];
-           const int remoteNStart = globalGrpTreeOffsetsFull[ibox];
-
-           const int remoteNBodies = host_float_as_int(globalGrpTreeCntSizeFull[remoteNStart].x);
-           const int remoteNNodes  = host_float_as_int(globalGrpTreeCntSizeFull[remoteNStart].y);
-
-           double bla;
-           const int resultTree = getLEToptQuickTreevsTree(
-               bufferStructTemp[tid],
-               &globalGrpTreeCntSizeFull[1+remoteNStart+remoteNBodies],
-               &globalGrpTreeCntSizeFull[1+remoteNStart+remoteNBodies+remoteNNodes],
-               &globalGrpTreeCntSizeFull[1+remoteNStart+remoteNBodies+remoteNNodes*2],
-               0, 1,            //Start at the root of remote boundary tree
-               groupLocalSize,  //Local boundarytree
-               groupLocalCntr,  //Local boundarytree
-               0, 1,            //start at the root of local boundary tree
-               remoteNNodes,
-               procId,
-               ibox,
-               bla);
-
-           if(resultTree == 0)
-           {
-             #pragma omp critical
-               countGroupsSuccess++;
-           }
-
-//           fprintf(stderr,"[Proc: %d  tid: %d ] The tree-to-tree test result for remote proc: %d Result: %d \n", procId, tid, ibox, resultTree);
-         } //while
-       } //omp section
-       sprintf(buff5,"ILETTIME-%d: nBoundaryOk: %d\n", procId, countGroupsSuccess);
-       devContext.writeLogEvent(buff5); //TODO DELETE
-       sprintf(buff5,"JLETTIME-%d: tBoundaryOk: %f\n", procId, get_time()-tStartNBoundaryOk);
-       devContext.writeLogEvent(buff5); //TODO DELETE
-     }//if 1
-
-     if(globalGrpTreeCntSizeFull) delete[] globalGrpTreeCntSizeFull;
-
-#if 1 //alltoallv version
-     {
-    mpiSync();
-    double tA2AGrpFullMid2 = get_time();
-
-     MPI_Allgather(&nGroupsFull,            sizeof(int), MPI_BYTE,
-                   (void*)&globalSizeArrayFull[0], sizeof(int), MPI_BYTE,
-                   mpiCommWorld); /* to globalSize Array */
-
-     std::vector<int> alltoallSend(nProcs);
-     std::vector<int> alltoallSendOff(nProcs, 0);
-     int runningOffsetFull = 0;
-     for (int i = 0; i < nProcs; i++)
-     {
-       globalGrpTreeCountFull[i]   = globalSizeArrayFull[i];
-       globalGrpTreeOffsetsFull[i] = runningOffsetFull;
-
-       displacementFull[i] = runningOffsetFull*sizeof(real4);
-
-       runningOffsetFull += globalSizeArrayFull[i];
-       globalSizeArrayFull[i] *= sizeof(real4);
-
-       alltoallSend[i] = globalSizeArrayFull[procId];
-     }
-
-     const int totalNumberOfGroupsFull = runningOffsetFull;  /*check if defined */
-     if(procId == 0) fprintf(stderr,"Proc: %d A2A Total number of groups full: %d  Time so far: %lg\n",
-              procId,runningOffsetFull, get_time()-tA2AGrpFullMid2);
-
-     globalGrpTreeCntSizeFull = new real4[totalNumberOfGroupsFull]; /* totalNumberOfGroups = 2*nGroups_recvd */
-
-
-     /* compute displacements for allgatherv */
-     MPI_Alltoallv(&fullBoundaryTree[0], &alltoallSend[0], &alltoallSendOff[0], MPI_BYTE,
-             &globalGrpTreeCntSizeFull[0], &globalSizeArrayFull[0], &displacementFull[0], MPI_BYTE,
-           mpiCommWorld);
-
-
-     double tA2AEndGrpFull = get_time(); //TODO delete
-
-     sprintf(buff5,"HLETTIME-%d: tA2AGrpFull: %lg\n", procId, tA2AEndGrpFull-tA2AGrpFullMid2);
-     devContext.writeLogEvent(buff5); //TODO DELETE
-     }
-#endif
-
-     if(globalGrpTreeCntSizeFull) delete[] globalGrpTreeCntSizeFull;
-
-
-
-    //Now we have a full working tree-structure composed only of the boundaries, lets print it to
-    //visualize it
-    if(0)
-    {
-       const int levelCountMax = localTree.n_nodes;
-       std::vector<int> currLevelVec, nextLevelVec;
-       currLevelVec.reserve(levelCountMax);
-       nextLevelVec.reserve(levelCountMax);
-       Swap<std::vector<int> > levelList(currLevelVec, nextLevelVec);
-
-       levelList.first().push_back(0);
-
-       char fileName[256];
-       ofstream nodeFile;
-
-       int depth = 0;
-       while (!levelList.first().empty())
-       {
-
-         sprintf(fileName, "BoundaryTreeStructure-%d-level-%d.txt", procId, depth);
-         nodeFile.open(fileName);
-         nodeFile << "NODES" << endl;
-         depth++;
-
-         const int csize = levelList.first().size();
-         for (int i = 0; i < csize; i++)
-         {
-           const uint   nodeIdx = levelList.first()[i];
-           const float4 centre  = groupCentre3[nodeIdx];
-           const float4 size    = groupSize3[nodeIdx];
-           const float nodeInfo_x = centre.w;
-           const uint  nodeInfo_y = host_float_as_int(size.w);
-
-           const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-           const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-
-           const bool lleaf = nodeInfo_x <= 0.0f;
-
-           //Write the properties to file
-           nodeFile << centre.x << "\t" << centre.y << "\t" << centre.z;
-           nodeFile << "\t"  << size.x << "\t" << size.y << "\t" << size.z << "\n";
-
-           //Check if it is an end point
-           if(nodeInfo_y == 0xFFFFFFFF) continue;
-
-           if (!lleaf)
-           {
-             //Add children to the next level list
-             for (int i = lchild; i < lchild + lnchild; i++)
-                      levelList.second().push_back(i);
-           }
-         }//end for
-
-         nodeFile.close();
-
-
-         levelList.swap();
-         levelList.second().clear();
-       }//end while
-    }//end section
-
-#endif
-
-#endif
-
-
-    template<typename T>
-    int getLEToptQuickCombinedTree(
-        std::vector<T> &LETBuffer,
-        GETLETBUFFERS &bufferStruct,
-        const int NCELLMAX,
-        const int NDEPTHMAX,
-        const real4 *nodeCentre,
-        const real4 *nodeSize,
-        const real4 *multipole,
-        const int cellBeg,
-        const int cellEnd,
-        const real4 *bodies,
-        const int nParticles,
-        const real4 *groupSizeInfo,
-        const real4 *groupCentreInfo,
-        const int groupBeg,
-        const int groupEnd,
-        const int nNodes,
-        const int procId,
-        const int ibox,
-        unsigned long long &nflops,
-        double &time)
-    {
-      double tStart = get_time2();
-
-      int depth = 0;
-
-      nflops = 0;
-
-      int nExportCell = 0;
-      int nExportPtcl = 0;
-      int nExportCellOffset = cellEnd;
-
-      const _v4sf*            bodiesV = (const _v4sf*)bodies;
-      const _v4sf*          nodeSizeV = (const _v4sf*)nodeSize;
-      const _v4sf*        nodeCentreV = (const _v4sf*)nodeCentre;
-      const _v4sf*         multipoleV = (const _v4sf*)multipole;
-      const _v4sf*   grpNodeSizeInfoV = (const _v4sf*)groupSizeInfo;
-      const _v4sf* grpNodeCenterInfoV = (const _v4sf*)groupCentreInfo;
-
-
-    #if 0 /* AVX */
-    #ifndef __AVX__
-    #error "AVX is not defined"
-    #endif
-      const int SIMDW  = 8;
-    #define AVXIMBH
-    #else
-      const int SIMDW  = 4;
-    #define SSEIMBH
-    #endif
-
-      bufferStruct.LETBuffer_node.clear();
-      bufferStruct.LETBuffer_ptcl.clear();
-      bufferStruct.currLevelVecUI4.clear();
-      bufferStruct.nextLevelVecUI4.clear();
-      bufferStruct.currGroupLevelVec.clear();
-      bufferStruct.nextGroupLevelVec.clear();
-      bufferStruct.groupSplitFlag.clear();
-
-      Swap<std::vector<uint4> > levelList(bufferStruct.currLevelVecUI4, bufferStruct.nextLevelVecUI4);
-      Swap<std::vector<int> > levelGroups(bufferStruct.currGroupLevelVec, bufferStruct.nextGroupLevelVec);
-
-      nExportCell += cellBeg;
-      for (int node = 0; node < cellBeg; node++)
-        bufferStruct.LETBuffer_node.push_back((int2){node, host_float_as_int(nodeSize[node].w)});
-
-      /* copy group info into current level buffer */
-      for (int group = groupBeg; group < groupEnd; group++)
-        levelGroups.first().push_back(group);
-
-      for (int cell = cellBeg; cell < cellEnd; cell++)
-        levelList.first().push_back((uint4){cell, 0, (int)levelGroups.first().size(),0});
-
-      double tPrep = get_time2();
-
-      while (!levelList.first().empty())
-      {
-        const int csize = levelList.first().size();
-        for (int i = 0; i < csize; i++)
-        {
-          /* play with criteria to fit what's best */
-          if (depth > NDEPTHMAX && nExportCell > NCELLMAX){
-    //        double tCalc = get_time2();
-            time = get_time2() - tStart;
-    //        fprintf(stderr,"[Proc: %d tid: %d ] getLETOptQuick exit ibox: %d depth: %d P: %d N: %d  Calc took: %lg Prepare: %lg Total: %lg \n",
-    //                        procId, omp_get_thread_num(), ibox, depth, nExportPtcl, nExportCell, tCalc-tPrep, tPrep - tStart, tCalc-tStart);
-            return -1;
-        }
-          if (nExportCell > NCELLMAX){
-            time = get_time2() - tStart;
-    //        double tCalc = get_time2();
-    //        fprintf(stderr,"[Proc: %d tid: %d ] getLETOptQuick exit ibox: %d depth: %d P: %d N: %d  Calc took: %lg Prepare: %lg Total: %lg \n",
-    //                        procId, omp_get_thread_num(), ibox, depth, nExportPtcl, nExportCell, tCalc-tPrep, tPrep - tStart, tCalc-tStart);
-            return -1;
-          }
-
-
-          const uint4       nodePacked = levelList.first()[i];
-          const uint  nodeIdx          = nodePacked.x;
-          const float nodeInfo_x       = nodeCentre[nodeIdx].w;
-          const uint  nodeInfo_y       = host_float_as_int(nodeSize[nodeIdx].w);
-
-          const _v4sf nodeCOM          = __builtin_ia32_vec_set_v4sf(multipoleV[nodeIdx*3], nodeInfo_x, 3);
-          const bool lleaf             = nodeInfo_x <= 0.0f;
-
-          const int groupBeg = nodePacked.y;
-          const int groupEnd = nodePacked.z;
-          nflops += 20*((groupEnd - groupBeg-1)/SIMDW+1)*SIMDW;
-
-          bufferStruct.groupSplitFlag.clear();
-          for (int ib = groupBeg; ib < groupEnd; ib += SIMDW)
-          {
-            _v4sf centre[SIMDW], size[SIMDW];
-            for (int laneIdx = 0; laneIdx < SIMDW; laneIdx++)
-            {
-              const int group = levelGroups.first()[std::min(ib+laneIdx, groupEnd-1)];
-              centre[laneIdx] = grpNodeCenterInfoV[group];
-              size  [laneIdx] =   grpNodeSizeInfoV[group];
-            }
-    #ifdef AVXIMBH
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box8a(nodeCOM, centre, size));
-    #else
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box4a(nodeCOM, centre, size));
-    #endif
-          }
-
-          const int groupNextBeg = levelGroups.second().size();
-          int split = false;
-          for (int idx = groupBeg; idx < groupEnd; idx++)
-          {
-            const bool gsplit = ((uint*)&bufferStruct.groupSplitFlag[0])[idx - groupBeg];
-
-            if (gsplit)
-            {
-              split = true;
-              const int group = levelGroups.first()[idx];
-              if (!lleaf)
-              {
-                //const bool gleaf = groupCentreInfo[group].w <= 0.0f; //This one does not go down leaves
-                const bool gleaf = groupCentreInfo[group].w == 0.0f; //Old tree This one goes up to including actual groups
-    //            const bool gleaf = groupCentreInfo[group].w == -1; //GPU-tree This one goes up to including actual groups
-                if (!gleaf)
-                {
-                  const int childinfoGrp  = ((uint4*)groupSizeInfo)[group].w;
-                  const int gchild  =   childinfoGrp & 0x0FFFFFFF;
-                  const int gnchild = ((childinfoGrp & 0xF0000000) >> 28) ;
-
-
-                  for (int i = gchild; i <= gchild+gnchild; i++) //old tree
-                  //for (int i = gchild; i < gchild+gnchild; i++) //GPU-tree TODO JB: I think this is the correct one, verify in treebuild code
-                  {
-                    levelGroups.second().push_back(i);
-                  }
-                }
-                else
-                  levelGroups.second().push_back(group);
-              }
-              else
-                break;
-            }
-          }
-
-          real4 size  = nodeSize[nodeIdx];
-          int sizew   = 0xFFFFFFFF;
-
-          if (split)
-          {
-            if (!lleaf)
-            {
-              const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-              const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-              sizew = (nExportCellOffset | (lnchild << LEAFBIT));
-              nExportCellOffset += lnchild;
-              for (int i = lchild; i < lchild + lnchild; i++)
-                levelList.second().push_back((uint4){i,groupNextBeg,(int)levelGroups.second().size()});
-            }
-            else
-            {
-              const int pfirst =    nodeInfo_y & BODYMASK;
-              const int np     = (((nodeInfo_y & INVBMASK) >> LEAFBIT)+1);
-              sizew = (nExportPtcl | ((np-1) << LEAFBIT));
-              for (int i = pfirst; i < pfirst+np; i++)
-                bufferStruct.LETBuffer_ptcl.push_back(i);
-              nExportPtcl += np;
-            }
-          }
-
-          bufferStruct.LETBuffer_node.push_back((int2){nodeIdx, sizew});
-          nExportCell++;
-
-        }
-        depth++;
-        levelList.swap();
-        levelList.second().clear();
-
-        levelGroups.swap();
-        levelGroups.second().clear();
-      }
-
-      double tCalc = get_time2();
-
-      assert((int)bufferStruct.LETBuffer_ptcl.size() == nExportPtcl);
-      assert((int)bufferStruct.LETBuffer_node.size() == nExportCell);
-
-      /* now copy data into LETBuffer */
-      {
-        _v4sf *vLETBuffer;
-        {
-          const size_t oldSize     = LETBuffer.size();
-          const size_t oldCapacity = LETBuffer.capacity();
-          LETBuffer.resize(oldSize + 1 + nExportPtcl + 5*nExportCell);
-          const size_t newCapacity = LETBuffer.capacity();
-          /* make sure memory is not reallocated */
-          assert(oldCapacity == newCapacity);
-
-          /* fill tree info */
-          real4 &data4 = *(real4*)(&LETBuffer[oldSize]);
-          data4.x      = host_int_as_float(nExportPtcl);
-          data4.y      = host_int_as_float(nExportCell);
-          data4.z      = host_int_as_float(cellBeg);
-          data4.w      = host_int_as_float(cellEnd);
-
-          //LOGF(stderr, "LET res for: %d  P: %d  N: %d old: %ld  Size in byte: %d\n",procId, nExportPtcl, nExportCell, oldSize, (int)(( 1 + nExportPtcl + 5*nExportCell)*sizeof(real4)));
-          vLETBuffer = (_v4sf*)(&LETBuffer[oldSize+1]);
-        }
-
-        int nStoreIdx     = nExportPtcl;
-        int multiStoreIdx = nStoreIdx + 2*nExportCell;
-        for (int i = 0; i < nExportPtcl; i++)
-        {
-          const int idx = bufferStruct.LETBuffer_ptcl[i];
-          vLETBuffer[i] = bodiesV[idx];
-        }
-        for (int i = 0; i < nExportCell; i++)
-        {
-          const int2 packed_idx = bufferStruct.LETBuffer_node[i];
-          const int idx = packed_idx.x;
-          const float sizew = host_int_as_float(packed_idx.y);
-          const _v4sf size = __builtin_ia32_vec_set_v4sf(nodeSizeV[idx], sizew, 3);
-
-          vLETBuffer[nStoreIdx+nExportCell] = nodeCentreV[idx];     /* centre */
-          vLETBuffer[nStoreIdx            ] = size;                 /*  size  */
-
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+0];  /* multipole com */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+1];  /* multipole q0 */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+2];  /* multipole q1 */
-          nStoreIdx++;
-        } //for
-      } //now copy data into LETBuffer
-
-      time = get_time2() - tStart;
-      double tEnd = get_time2();
-
-     LOGF(stderr,"getLETOptQuickCombinedTree P: %d N: %d  Calc took: %lg Prepare: %lg Copy: %lg Total: %lg \n",nExportPtcl, nExportCell, tCalc-tStart, tPrep - tStart, tEnd-tCalc, tEnd-tStart);
-    //  fprintf(stderr,"[Proc: %d ] getLETOptQuick P: %d N: %d  Calc took: %lg Prepare: %lg (calc: %lg ) Copy: %lg Total: %lg \n",
-    //    procId, nExportPtcl, nExportCell, tCalc-tStart, tPrep - tStart, tCalc-tPrep,  tEnd-tCalc, tEnd-tStart);
-
-      return  1 + nExportPtcl + 5*nExportCell;
-    }
-
-
-    template<typename T>
-    int getLEToptQuick(
-        std::vector<T> &LETBuffer,
-        GETLETBUFFERS &bufferStruct,
-        const int NCELLMAX,
-        const int NDEPTHMAX,
-        const real4 *nodeCentre,
-        const real4 *nodeSize,
-        const real4 *multipole,
-        const int cellBeg,
-        const int cellEnd,
-        const real4 *bodies,
-        const int nParticles,
-        const real4 *groupSizeInfo,
-        const real4 *groupCentreInfo,
-        const int groupBeg,
-        const int groupEnd,
-        const int nNodes,
-        const int procId,
-        const int ibox,
-        unsigned long long &nflops,
-        double &time)
-    {
-      double tStart = get_time2();
-
-      int depth = 0;
-
-      nflops = 0;
-
-      int nExportCell = 0;
-      int nExportPtcl = 0;
-      int nExportCellOffset = cellEnd;
-
-      const _v4sf*            bodiesV = (const _v4sf*)bodies;
-      const _v4sf*          nodeSizeV = (const _v4sf*)nodeSize;
-      const _v4sf*        nodeCentreV = (const _v4sf*)nodeCentre;
-      const _v4sf*         multipoleV = (const _v4sf*)multipole;
-      const _v4sf*   grpNodeSizeInfoV = (const _v4sf*)groupSizeInfo;
-      const _v4sf* grpNodeCenterInfoV = (const _v4sf*)groupCentreInfo;
-
-
-    #if 0 /* AVX */
-    #ifndef __AVX__
-    #error "AVX is not defined"
-    #endif
-      const int SIMDW  = 8;
-    #define AVXIMBH
-    #else
-      const int SIMDW  = 4;
-    #define SSEIMBH
-    #endif
-
-      bufferStruct.LETBuffer_node.clear();
-      bufferStruct.LETBuffer_ptcl.clear();
-      bufferStruct.currLevelVecUI4.clear();
-      bufferStruct.nextLevelVecUI4.clear();
-      bufferStruct.currGroupLevelVec.clear();
-      bufferStruct.nextGroupLevelVec.clear();
-      bufferStruct.groupSplitFlag.clear();
-
-      Swap<std::vector<uint4> > levelList(bufferStruct.currLevelVecUI4, bufferStruct.nextLevelVecUI4);
-      Swap<std::vector<int> > levelGroups(bufferStruct.currGroupLevelVec, bufferStruct.nextGroupLevelVec);
-
-      nExportCell += cellBeg;
-      for (int node = 0; node < cellBeg; node++)
-        bufferStruct.LETBuffer_node.push_back((int2){node, host_float_as_int(nodeSize[node].w)});
-
-      /* copy group info into current level buffer */
-      for (int group = groupBeg; group < groupEnd; group++)
-        levelGroups.first().push_back(group);
-
-      for (int cell = cellBeg; cell < cellEnd; cell++)
-        levelList.first().push_back((uint4){cell, 0, (int)levelGroups.first().size(),0});
-
-      double tPrep = get_time2();
-
-      while (!levelList.first().empty())
-      {
-        const int csize = levelList.first().size();
-        for (int i = 0; i < csize; i++)
-        {
-          /* play with criteria to fit what's best */
-          if (depth > NDEPTHMAX && nExportCell > NCELLMAX){
-    //        double tCalc = get_time2();
-            time = get_time2() - tStart;
-    //        fprintf(stderr,"[Proc: %d tid: %d ] getLETOptQuick exit ibox: %d depth: %d P: %d N: %d  Calc took: %lg Prepare: %lg Total: %lg \n",
-    //                        procId, omp_get_thread_num(), ibox, depth, nExportPtcl, nExportCell, tCalc-tPrep, tPrep - tStart, tCalc-tStart);
-            return -1;
-        }
-          if (nExportCell > NCELLMAX){
-            time = get_time2() - tStart;
-    //        double tCalc = get_time2();
-    //        fprintf(stderr,"[Proc: %d tid: %d ] getLETOptQuick exit ibox: %d depth: %d P: %d N: %d  Calc took: %lg Prepare: %lg Total: %lg \n",
-    //                        procId, omp_get_thread_num(), ibox, depth, nExportPtcl, nExportCell, tCalc-tPrep, tPrep - tStart, tCalc-tStart);
-            return -1;
-          }
-
-
-          const uint4       nodePacked = levelList.first()[i];
-          const uint  nodeIdx          = nodePacked.x;
-          const float nodeInfo_x       = nodeCentre[nodeIdx].w;
-          const uint  nodeInfo_y       = host_float_as_int(nodeSize[nodeIdx].w);
-
-          const _v4sf nodeCOM          = __builtin_ia32_vec_set_v4sf(multipoleV[nodeIdx*3], nodeInfo_x, 3);
-          const bool lleaf             = nodeInfo_x <= 0.0f;
-
-          const int groupBeg = nodePacked.y;
-          const int groupEnd = nodePacked.z;
-          nflops += 20*((groupEnd - groupBeg-1)/SIMDW+1)*SIMDW;
-
-          bufferStruct.groupSplitFlag.clear();
-          for (int ib = groupBeg; ib < groupEnd; ib += SIMDW)
-          {
-            _v4sf centre[SIMDW], size[SIMDW];
-            for (int laneIdx = 0; laneIdx < SIMDW; laneIdx++)
-            {
-              const int group = levelGroups.first()[std::min(ib+laneIdx, groupEnd-1)];
-              centre[laneIdx] = grpNodeCenterInfoV[group];
-              size  [laneIdx] =   grpNodeSizeInfoV[group];
-            }
-    #ifdef AVXIMBH
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box8a(nodeCOM, centre, size));
-    #else
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box4a(nodeCOM, centre, size));
-    #endif
-          }
-
-          const int groupNextBeg = levelGroups.second().size();
-          int split = false;
-          for (int idx = groupBeg; idx < groupEnd; idx++)
-          {
-            const bool gsplit = ((uint*)&bufferStruct.groupSplitFlag[0])[idx - groupBeg];
-
-            if (gsplit)
-            {
-              split = true;
-              const int group = levelGroups.first()[idx];
-              if (!lleaf)
-              {
-                //const bool gleaf = groupCentreInfo[group].w <= 0.0f; //This one does not go down leaves
-    //            const bool gleaf = groupCentreInfo[group].w == 0.0f; //Old tree This one goes up to including actual groups
-                const bool gleaf = groupCentreInfo[group].w == -1; //GPU-tree This one goes up to including actual groups
-                if (!gleaf)
-                {
-                  const int childinfoGrp  = ((uint4*)groupSizeInfo)[group].w;
-                  const int gchild  =   childinfoGrp & 0x0FFFFFFF;
-                  const int gnchild = ((childinfoGrp & 0xF0000000) >> 28) ;
-
-
-                  //for (int i = gchild; i <= gchild+gnchild; i++) //old tree
-                  for (int i = gchild; i < gchild+gnchild; i++) //GPU-tree TODO JB: I think this is the correct one, verify in treebuild code
-                  {
-                    levelGroups.second().push_back(i);
-                  }
-                }
-                else
-                  levelGroups.second().push_back(group);
-              }
-              else
-                break;
-            }
-          }
-
-          real4 size  = nodeSize[nodeIdx];
-          int sizew   = 0xFFFFFFFF;
-
-          if (split)
-          {
-            if (!lleaf)
-            {
-              const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-              const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-              sizew = (nExportCellOffset | (lnchild << LEAFBIT));
-              nExportCellOffset += lnchild;
-              for (int i = lchild; i < lchild + lnchild; i++)
-                levelList.second().push_back((uint4){i,groupNextBeg,(int)levelGroups.second().size()});
-            }
-            else
-            {
-              const int pfirst =    nodeInfo_y & BODYMASK;
-              const int np     = (((nodeInfo_y & INVBMASK) >> LEAFBIT)+1);
-              sizew = (nExportPtcl | ((np-1) << LEAFBIT));
-              for (int i = pfirst; i < pfirst+np; i++)
-                bufferStruct.LETBuffer_ptcl.push_back(i);
-              nExportPtcl += np;
-            }
-          }
-
-          bufferStruct.LETBuffer_node.push_back((int2){nodeIdx, sizew});
-          nExportCell++;
-
-        }
-        depth++;
-        levelList.swap();
-        levelList.second().clear();
-
-        levelGroups.swap();
-        levelGroups.second().clear();
-      }
-
-      double tCalc = get_time2();
-
-      assert((int)bufferStruct.LETBuffer_ptcl.size() == nExportPtcl);
-      assert((int)bufferStruct.LETBuffer_node.size() == nExportCell);
-
-      /* now copy data into LETBuffer */
-      {
-        _v4sf *vLETBuffer;
-        {
-          const size_t oldSize     = LETBuffer.size();
-          const size_t oldCapacity = LETBuffer.capacity();
-          LETBuffer.resize(oldSize + 1 + nExportPtcl + 5*nExportCell);
-          const size_t newCapacity = LETBuffer.capacity();
-          /* make sure memory is not reallocated */
-          assert(oldCapacity == newCapacity);
-
-          /* fill tree info */
-          real4 &data4 = *(real4*)(&LETBuffer[oldSize]);
-          data4.x      = host_int_as_float(nExportPtcl);
-          data4.y      = host_int_as_float(nExportCell);
-          data4.z      = host_int_as_float(cellBeg);
-          data4.w      = host_int_as_float(cellEnd);
-
-          //LOGF(stderr, "LET res for: %d  P: %d  N: %d old: %ld  Size in byte: %d\n",procId, nExportPtcl, nExportCell, oldSize, (int)(( 1 + nExportPtcl + 5*nExportCell)*sizeof(real4)));
-          vLETBuffer = (_v4sf*)(&LETBuffer[oldSize+1]);
-        }
-
-        int nStoreIdx     = nExportPtcl;
-        int multiStoreIdx = nStoreIdx + 2*nExportCell;
-        for (int i = 0; i < nExportPtcl; i++)
-        {
-          const int idx = bufferStruct.LETBuffer_ptcl[i];
-          vLETBuffer[i] = bodiesV[idx];
-        }
-        for (int i = 0; i < nExportCell; i++)
-        {
-          const int2 packed_idx = bufferStruct.LETBuffer_node[i];
-          const int idx = packed_idx.x;
-          const float sizew = host_int_as_float(packed_idx.y);
-          const _v4sf size = __builtin_ia32_vec_set_v4sf(nodeSizeV[idx], sizew, 3);
-
-          vLETBuffer[nStoreIdx+nExportCell] = nodeCentreV[idx];     /* centre */
-          vLETBuffer[nStoreIdx            ] = size;                 /*  size  */
-
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+0];  /* multipole com */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+1];  /* multipole q0 */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+2];  /* multipole q1 */
-          nStoreIdx++;
-        } //for
-      } //now copy data into LETBuffer
-
-      time = get_time2() - tStart;
-      double tEnd = get_time2();
-
-    // LOGF(stderr,"getLETOptQuick P: %d N: %d  Calc took: %lg Prepare: %lg Copy: %lg Total: %lg \n",nExportPtcl, nExportCell, tCalc-tStart, tPrep - tStart, tEnd-tCalc, tEnd-tStart);
-    //  fprintf(stderr,"[Proc: %d ] getLETOptQuick P: %d N: %d  Calc took: %lg Prepare: %lg (calc: %lg ) Copy: %lg Total: %lg \n",
-    //    procId, nExportPtcl, nExportCell, tCalc-tStart, tPrep - tStart, tCalc-tPrep,  tEnd-tCalc, tEnd-tStart);
-
-      return  1 + nExportPtcl + 5*nExportCell;
-    }
-
-
-    template<typename T>
-    int getLEToptQuickCombinedTreeBoundCheck(
-        std::vector<T> &LETBuffer,
-        GETLETBUFFERS &bufferStruct,
-        std::vector<int> &boundaryIds,
-        const int NCELLMAX,
-        const int NDEPTHMAX,
-        const real4 *nodeCentre,
-        const real4 *nodeSize,
-        const real4 *multipole,
-        const int cellBeg,
-        const int cellEnd,
-        const real4 *bodies,
-        const int nParticles,
-        const real4 *groupSizeInfo,
-        const real4 *groupCentreInfo,
-        const int groupBeg,
-        const int groupEnd,
-        const int nNodes,
-        const int procId,
-        const int ibox,
-        unsigned long long &nflops,
-        double &time)
-    {
-      double tStart = get_time2();
-
-      int depth = 0;
-
-      nflops = 0;
-
-      double haveToSend = false;
-
-      int nExportCell = 0;
-      int nExportPtcl = 0;
-      int nExportCellOffset = cellEnd;
-
-      const _v4sf*            bodiesV = (const _v4sf*)bodies;
-      const _v4sf*          nodeSizeV = (const _v4sf*)nodeSize;
-      const _v4sf*        nodeCentreV = (const _v4sf*)nodeCentre;
-      const _v4sf*         multipoleV = (const _v4sf*)multipole;
-      const _v4sf*   grpNodeSizeInfoV = (const _v4sf*)groupSizeInfo;
-      const _v4sf* grpNodeCenterInfoV = (const _v4sf*)groupCentreInfo;
-
-
-    #if 0 /* AVX */
-    #ifndef __AVX__
-    #error "AVX is not defined"
-    #endif
-      const int SIMDW  = 8;
-    #define AVXIMBH
-    #else
-      const int SIMDW  = 4;
-    #define SSEIMBH
-    #endif
-
-      bufferStruct.LETBuffer_node.clear();
-      bufferStruct.LETBuffer_ptcl.clear();
-      bufferStruct.currLevelVecUI4.clear();
-      bufferStruct.nextLevelVecUI4.clear();
-      bufferStruct.currGroupLevelVec.clear();
-      bufferStruct.nextGroupLevelVec.clear();
-      bufferStruct.groupSplitFlag.clear();
-
-      Swap<std::vector<uint4> > levelList(bufferStruct.currLevelVecUI4, bufferStruct.nextLevelVecUI4);
-      Swap<std::vector<int> > levelGroups(bufferStruct.currGroupLevelVec, bufferStruct.nextGroupLevelVec);
-
-      nExportCell += cellBeg;
-      for (int node = 0; node < cellBeg; node++)
-        bufferStruct.LETBuffer_node.push_back((int2){node, host_float_as_int(nodeSize[node].w)});
-
-      /* copy group info into current level buffer */
-      for (int group = groupBeg; group < groupEnd; group++)
-        levelGroups.first().push_back(group);
-
-      for (int cell = cellBeg; cell < cellEnd; cell++)
-        levelList.first().push_back((uint4){cell, 0, (int)levelGroups.first().size(),0});
-
-      double tPrep = get_time2();
-
-      while (!levelList.first().empty())
-      {
-        const int csize = levelList.first().size();
-        for (int i = 0; i < csize; i++)
-        {
-          /* play with criteria to fit what's best */
-          if (depth > NDEPTHMAX && nExportCell > NCELLMAX){
-    //        double tCalc = get_time2();
-            time = get_time2() - tStart;
-    //        fprintf(stderr,"[Proc: %d tid: %d ] getLETOptQuick exit ibox: %d depth: %d P: %d N: %d  Calc took: %lg Prepare: %lg Total: %lg \n",
-    //                        procId, omp_get_thread_num(), ibox, depth, nExportPtcl, nExportCell, tCalc-tPrep, tPrep - tStart, tCalc-tStart);
-            return -1;
-        }
-          if (nExportCell > NCELLMAX){
-            time = get_time2() - tStart;
-    //        double tCalc = get_time2();
-    //        fprintf(stderr,"[Proc: %d tid: %d ] getLETOptQuick exit ibox: %d depth: %d P: %d N: %d  Calc took: %lg Prepare: %lg Total: %lg \n",
-    //                        procId, omp_get_thread_num(), ibox, depth, nExportPtcl, nExportCell, tCalc-tPrep, tPrep - tStart, tCalc-tStart);
-            return -1;
-          }
-
-
-          const uint4       nodePacked = levelList.first()[i];
-          const uint  nodeIdx          = nodePacked.x;
-          const float nodeInfo_x       = nodeCentre[nodeIdx].w;
-          const uint  nodeInfo_y       = host_float_as_int(nodeSize[nodeIdx].w);
-
-          const _v4sf nodeCOM          = __builtin_ia32_vec_set_v4sf(multipoleV[nodeIdx*3], nodeInfo_x, 3);
-          const bool lleaf             = nodeInfo_x <= 0.0f;
-
-          const int groupBeg = nodePacked.y;
-          const int groupEnd = nodePacked.z;
-          nflops += 20*((groupEnd - groupBeg-1)/SIMDW+1)*SIMDW;
-
-          bufferStruct.groupSplitFlag.clear();
-          for (int ib = groupBeg; ib < groupEnd; ib += SIMDW)
-          {
-            _v4sf centre[SIMDW], size[SIMDW];
-            for (int laneIdx = 0; laneIdx < SIMDW; laneIdx++)
-            {
-              const int group = levelGroups.first()[std::min(ib+laneIdx, groupEnd-1)];
-              centre[laneIdx] = grpNodeCenterInfoV[group];
-              size  [laneIdx] =   grpNodeSizeInfoV[group];
-            }
-    #ifdef AVXIMBH
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box8a(nodeCOM, centre, size));
-    #else
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box4a(nodeCOM, centre, size));
-    #endif
-          }
-
-          const int groupNextBeg = levelGroups.second().size();
-          int split = false;
-          for (int idx = groupBeg; idx < groupEnd; idx++)
-          {
-            const bool gsplit = ((uint*)&bufferStruct.groupSplitFlag[0])[idx - groupBeg];
-
-            if (gsplit)
-            {
-              split = true;
-              const int group = levelGroups.first()[idx];
-              if (!lleaf)
-              {
-                //const bool gleaf = groupCentreInfo[group].w <= 0.0f; //This one does not go down leaves
-                const bool gleaf = groupCentreInfo[group].w == 0.0f; //Old tree This one goes up to including actual groups
-    //            const bool gleaf = groupCentreInfo[group].w == -1; //GPU-tree This one goes up to including actual groups
-                if (!gleaf)
-                {
-                  const int childinfoGrp  = ((uint4*)groupSizeInfo)[group].w;
-                  const int gchild  =   childinfoGrp & 0x0FFFFFFF;
-                  const int gnchild = ((childinfoGrp & 0xF0000000) >> 28) ;
-
-
-                  for (int i = gchild; i <= gchild+gnchild; i++) //old tree
-                  //for (int i = gchild; i < gchild+gnchild; i++) //GPU-tree TODO JB: I think this is the correct one, verify in treebuild code
-                  {
-                    levelGroups.second().push_back(i);
-                  }
-                }
-                else
-                  levelGroups.second().push_back(group);
-              }
-              else
-                break;
-            }
-          }
-
-          real4 size  = nodeSize[nodeIdx];
-          int sizew   = 0xFFFFFFFF;
-
-          if (split)
-          {
-            if (!lleaf)
-            {
-              const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-              const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-              sizew = (nExportCellOffset | (lnchild << LEAFBIT));
-              nExportCellOffset += lnchild;
-              for (int i = lchild; i < lchild + lnchild; i++)
-                levelList.second().push_back((uint4){i,groupNextBeg,(int)levelGroups.second().size()});
-            }
-            else
-            {
-              const int pfirst =    nodeInfo_y & BODYMASK;
-              const int np     = (((nodeInfo_y & INVBMASK) >> LEAFBIT)+1);
-              sizew = (nExportPtcl | ((np-1) << LEAFBIT));
-              for (int i = pfirst; i < pfirst+np; i++)
-                bufferStruct.LETBuffer_ptcl.push_back(i);
-              nExportPtcl += np;
-            }
-          }
-
-
-          //Check if this node is included in the boundary-structure. If it is not in
-          //the boundary structure it means we have to send this tree-structure. Only check
-          //if we are undecided
-          if(!haveToSend)
-          {
-            if(boundaryIds[nodeIdx] != 1) haveToSend = true;
-          }
-
-          bufferStruct.LETBuffer_node.push_back((int2){nodeIdx, sizew});
-          nExportCell++;
-
-        }
-        depth++;
-        levelList.swap();
-        levelList.second().clear();
-
-        levelGroups.swap();
-        levelGroups.second().clear();
-      }
-
-      double tCalc = get_time2();
-
-      if(!haveToSend)
-      {
-        //All the found boxes are already part of the boundary structure
-        //so lets ignore them. We don't have to send this tree
-        return -2;
-      }
-
-    //  Hier gebleven. Dit afmaken, kijken hoe lang/snel het is om onze boundaries
-    //  test uit te voeren. Gebruik de test code voor het maken/testen van de code
-    //  test_localExtraBound -> de roots van de boundaries samen voegen in 1 structure
-    //  dan die roots koppelen aan elkaar en die volledige tree testen tegen onze eigen
-    //  tree of boundary tree. Belangrijk is dat het sneller is dan quickCheck
-
-
-
-      assert((int)bufferStruct.LETBuffer_ptcl.size() == nExportPtcl);
-      assert((int)bufferStruct.LETBuffer_node.size() == nExportCell);
-
-      /* now copy data into LETBuffer */
-      {
-        _v4sf *vLETBuffer;
-        {
-          const size_t oldSize     = LETBuffer.size();
-          const size_t oldCapacity = LETBuffer.capacity();
-          LETBuffer.resize(oldSize + 1 + nExportPtcl + 5*nExportCell);
-          const size_t newCapacity = LETBuffer.capacity();
-          /* make sure memory is not reallocated */
-          assert(oldCapacity == newCapacity);
-
-          /* fill tree info */
-          real4 &data4 = *(real4*)(&LETBuffer[oldSize]);
-          data4.x      = host_int_as_float(nExportPtcl);
-          data4.y      = host_int_as_float(nExportCell);
-          data4.z      = host_int_as_float(cellBeg);
-          data4.w      = host_int_as_float(cellEnd);
-
-          //LOGF(stderr, "LET res for: %d  P: %d  N: %d old: %ld  Size in byte: %d\n",procId, nExportPtcl, nExportCell, oldSize, (int)(( 1 + nExportPtcl + 5*nExportCell)*sizeof(real4)));
-          vLETBuffer = (_v4sf*)(&LETBuffer[oldSize+1]);
-        }
-
-        int nStoreIdx     = nExportPtcl;
-        int multiStoreIdx = nStoreIdx + 2*nExportCell;
-        for (int i = 0; i < nExportPtcl; i++)
-        {
-          const int idx = bufferStruct.LETBuffer_ptcl[i];
-          vLETBuffer[i] = bodiesV[idx];
-        }
-        for (int i = 0; i < nExportCell; i++)
-        {
-          const int2 packed_idx = bufferStruct.LETBuffer_node[i];
-          const int idx = packed_idx.x;
-          const float sizew = host_int_as_float(packed_idx.y);
-          const _v4sf size = __builtin_ia32_vec_set_v4sf(nodeSizeV[idx], sizew, 3);
-
-          vLETBuffer[nStoreIdx+nExportCell] = nodeCentreV[idx];     /* centre */
-          vLETBuffer[nStoreIdx            ] = size;                 /*  size  */
-
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+0];  /* multipole com */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+1];  /* multipole q0 */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+2];  /* multipole q1 */
-          nStoreIdx++;
-        } //for
-      } //now copy data into LETBuffer
-
-      time = get_time2() - tStart;
-      double tEnd = get_time2();
-
-     LOGF(stderr,"getLETOptQuickCombinedTree P: %d N: %d  Calc took: %lg Prepare: %lg Copy: %lg Total: %lg \n",nExportPtcl, nExportCell, tCalc-tStart, tPrep - tStart, tEnd-tCalc, tEnd-tStart);
-    //  fprintf(stderr,"[Proc: %d ] getLETOptQuick P: %d N: %d  Calc took: %lg Prepare: %lg (calc: %lg ) Copy: %lg Total: %lg \n",
-    //    procId, nExportPtcl, nExportCell, tCalc-tStart, tPrep - tStart, tCalc-tPrep,  tEnd-tCalc, tEnd-tStart);
-
-      return  1 + nExportPtcl + 5*nExportCell;
-    }
-
-
-
-    int3 getLETopt(
-        GETLETBUFFERS &bufferStruct,
-        real4 **LETBuffer_ptr,
-        const real4 *nodeCentre,
-        const real4 *nodeSize,
-        const real4 *multipole,
-        const int cellBeg,
-        const int cellEnd,
-        const real4 *bodies,
-        const int nParticles,
-        const real4 *groupSizeInfo,
-        const real4 *groupCentreInfo,
-        const int groupBeg,
-        const int groupEnd,
-        const int nNodes,
-        unsigned long long &nflops)
-    {
-      bufferStruct.LETBuffer_node.clear();
-      bufferStruct.LETBuffer_ptcl.clear();
-      bufferStruct.currLevelVecUI4.clear();
-      bufferStruct.nextLevelVecUI4.clear();
-      bufferStruct.currGroupLevelVec.clear();
-      bufferStruct.nextGroupLevelVec.clear();
-      bufferStruct.groupSplitFlag.clear();
-
-      nflops = 0;
-
-      int nExportCell = 0;
-      int nExportPtcl = 0;
-      int nExportCellOffset = cellEnd;
-
-      const _v4sf*            bodiesV = (const _v4sf*)bodies;
-      const _v4sf*          nodeSizeV = (const _v4sf*)nodeSize;
-      const _v4sf*        nodeCentreV = (const _v4sf*)nodeCentre;
-      const _v4sf*         multipoleV = (const _v4sf*)multipole;
-      const _v4sf*   grpNodeSizeInfoV = (const _v4sf*)groupSizeInfo;
-      const _v4sf* grpNodeCenterInfoV = (const _v4sf*)groupCentreInfo;
-
-
-    #if 0 /* AVX */
-    #ifndef __AVX__
-    #error "AVX is not defined"
-    #endif
-      const int SIMDW  = 8;
-    #define AVXIMBH
-    #else
-      const int SIMDW  = 4;
-    #define SSEIMBH
-    #endif
-
-
-      Swap<std::vector<uint4> > levelList(bufferStruct.currLevelVecUI4, bufferStruct.nextLevelVecUI4);
-      Swap<std::vector<int> > levelGroups(bufferStruct.currGroupLevelVec, bufferStruct.nextGroupLevelVec);
-
-      nExportCell += cellBeg;
-      for (int node = 0; node < cellBeg; node++)
-        bufferStruct.LETBuffer_node.push_back((int2){node, host_float_as_int(nodeSize[node].w)});
-
-
-
-      /* copy group info into current level buffer */
-      for (int group = groupBeg; group < groupEnd; group++)
-        levelGroups.first().push_back(group);
-
-      for (int cell = cellBeg; cell < cellEnd; cell++)
-        levelList.first().push_back((uint4){cell, 0, (int)levelGroups.first().size(),0});
-
-      int depth = 0;
-      while (!levelList.first().empty())
-      {
-        const int csize = levelList.first().size();
-        for (int i = 0; i < csize; i++)
-        {
-          const uint4       nodePacked = levelList.first()[i];
-
-          const uint  nodeIdx  = nodePacked.x;
-          const float nodeInfo_x = nodeCentre[nodeIdx].w;
-          const uint  nodeInfo_y = host_float_as_int(nodeSize[nodeIdx].w);
-
-          const _v4sf nodeCOM  = __builtin_ia32_vec_set_v4sf(multipoleV[nodeIdx*3], nodeInfo_x, 3);
-          const bool lleaf = nodeInfo_x <= 0.0f;
-
-          const int groupBeg = nodePacked.y;
-          const int groupEnd = nodePacked.z;
-          nflops += 20*((groupEnd - groupBeg-1)/SIMDW+1)*SIMDW;
-
-          bufferStruct.groupSplitFlag.clear();
-          for (int ib = groupBeg; ib < groupEnd; ib += SIMDW)
-          {
-            _v4sf centre[SIMDW], size[SIMDW];
-            for (int laneIdx = 0; laneIdx < SIMDW; laneIdx++)
-            {
-              const int group = levelGroups.first()[std::min(ib+laneIdx, groupEnd-1)];
-              centre[laneIdx] = grpNodeCenterInfoV[group];
-              size  [laneIdx] =   grpNodeSizeInfoV[group];
-            }
-    #ifdef AVXIMBH
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box8a(nodeCOM, centre, size));
-    #else
-            bufferStruct.groupSplitFlag.push_back(split_node_grav_impbh_box4a(nodeCOM, centre, size));
-    #endif
-          }
-
-          const int groupNextBeg = levelGroups.second().size();
-          int split = false;
-          for (int idx = groupBeg; idx < groupEnd; idx++)
-          {
-            const bool gsplit = ((uint*)&bufferStruct.groupSplitFlag[0])[idx - groupBeg];
-            if (gsplit)
-            {
-              split = true;
-              const int group = levelGroups.first()[idx];
-              if (!lleaf)
-              {
-                const bool gleaf = groupCentreInfo[group].w <= 0.0f;
-                if (!gleaf)
-                {
-                  const int childinfoGrp  = ((uint4*)groupSizeInfo)[group].w;
-                  const int gchild  =   childinfoGrp & 0x0FFFFFFF;
-                  const int gnchild = ((childinfoGrp & 0xF0000000) >> 28) ;
-                  for (int i = gchild; i <= gchild+gnchild; i++)
-                    levelGroups.second().push_back(i);
-                }
-                else
-                  levelGroups.second().push_back(group);
-              }
-              else
-                break;
-            }
-          }
-
-          real4 size  = nodeSize[nodeIdx];
-          int sizew = 0xFFFFFFFF;
-
-          if (split)
-          {
-            if (!lleaf)
-            {
-              const int lchild  =    nodeInfo_y & 0x0FFFFFFF;            //Index to the first child of the node
-              const int lnchild = (((nodeInfo_y & 0xF0000000) >> 28)) ;  //The number of children this node has
-              sizew = (nExportCellOffset | (lnchild << LEAFBIT));
-              nExportCellOffset += lnchild;
-              for (int i = lchild; i < lchild + lnchild; i++)
-                levelList.second().push_back((uint4){i,groupNextBeg,(int)levelGroups.second().size()});
-            }
-            else
-            {
-              const int pfirst =    nodeInfo_y & BODYMASK;
-              const int np     = (((nodeInfo_y & INVBMASK) >> LEAFBIT)+1);
-              sizew = (nExportPtcl | ((np-1) << LEAFBIT));
-              for (int i = pfirst; i < pfirst+np; i++)
-                bufferStruct.LETBuffer_ptcl.push_back(i);
-              nExportPtcl += np;
-            }
-          }
-
-          bufferStruct.LETBuffer_node.push_back((int2){nodeIdx, sizew});
-          nExportCell++;
-        }
-
-        depth++;
-
-        levelList.swap();
-        levelList.second().clear();
-
-        levelGroups.swap();
-        levelGroups.second().clear();
-      }
-
-      assert((int)bufferStruct.LETBuffer_ptcl.size() == nExportPtcl);
-      assert((int)bufferStruct.LETBuffer_node.size() == nExportCell);
-
-      /* now copy data into LETBuffer */
-      {
-        //LETBuffer.resize(nExportPtcl + 5*nExportCell);
-    #pragma omp critical //Malloc seems to be not so thread safe..
-        *LETBuffer_ptr = (real4*)malloc(sizeof(real4)*(1+ nExportPtcl + 5*nExportCell));
-        real4 *LETBuffer = *LETBuffer_ptr;
-        _v4sf *vLETBuffer      = (_v4sf*)(&LETBuffer[1]);
-        //_v4sf *vLETBuffer      = (_v4sf*)&LETBuffer     [0];
-
-        int nStoreIdx = nExportPtcl;
-        int multiStoreIdx = nStoreIdx + 2*nExportCell;
-        for (int i = 0; i < nExportPtcl; i++)
-        {
-          const int idx = bufferStruct.LETBuffer_ptcl[i];
-          vLETBuffer[i] = bodiesV[idx];
-        }
-        for (int i = 0; i < nExportCell; i++)
-        {
-          const int2 packed_idx = bufferStruct.LETBuffer_node[i];
-          const int idx = packed_idx.x;
-          const float sizew = host_int_as_float(packed_idx.y);
-          const _v4sf size = __builtin_ia32_vec_set_v4sf(nodeSizeV[idx], sizew, 3);
-
-          //       vLETBuffer[nStoreIdx            ] = nodeCentreV[idx];     /* centre */
-          //       vLETBuffer[nStoreIdx+nExportCell] = size;                 /*  size  */
-
-          vLETBuffer[nStoreIdx+nExportCell] = nodeCentreV[idx];     /* centre */
-          vLETBuffer[nStoreIdx            ] = size;                 /*  size  */
-
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+0];  /* multipole.x */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+1];  /* multipole.x */
-          vLETBuffer[multiStoreIdx++      ] = multipoleV[3*idx+2];  /* multipole.x */
-          nStoreIdx++;
-        }
-      }
-
-      return (int3){nExportCell, nExportPtcl, depth};
-    }
-
-
-
-    Domain check. Compare host and device
-      validList2.d2h();
-
-      localTree.bodies_key.d2h();
-      for(int ib=0;ib<nProcs;ib++)
-      {
-        int ibox          = ib;
-        for(int i=0; i<localTree.n;i++)
-        {
-          uint4 lowerBoundary = localTree.parallelBoundaries[ibox];
-          uint4 upperBoundary = localTree.parallelBoundaries[ibox+1];
-
-          uint4 key  = localTree.bodies_key[i];
-          int bottom = cmp_uint4(key, lowerBoundary);
-          int top    = cmp_uint4(key, upperBoundary);
-
-          if(bottom >= 0 && top < 0)
-          {
-            validList2[i].x = validList2[i].x & 0x0FFFFFF;
-
-            if( validList2[i].x != ibox)
-              if(!( !(validList2[i].x >> 31) && ibox == procId))
-                LOGF(stderr,"Particle: %d  GPU says: %d Host says: %d \n", i, validList2[i].x, ibox);
-          }
-        }//for i
-      }//for ib
-
-#endif //if 1/0
-
-#if 0
-
-
-//Exchange particles with other processes
-int octree::gpu_exchange_particles_with_overflow_check_SFC(tree_structure &tree,
-    bodyStruct *particlesToSend,
-    my_dev::dev_mem<uint> &extractList,
-
-    int nToSend)
-{
-#ifdef USE_MPI
-
-  double tStart = get_time();
-
-  int myid      = procId;
-  int nproc     = nProcs;
-  int iloc      = 0;
-  int nbody     = nToSend;
-
-
-  bodyStruct  tmpp;
-
-  int *firstloc    = &exchangePartBuffer[0*nProcs];                //Size nProcs+1
-  int *nparticles  = &exchangePartBuffer[1*(nProcs+1)];            //Size nProcs+1
-  int *nsendDispls = &exchangePartBuffer[2*(nProcs+1)];            //Size nProcs+1
-  int *nrecvDispls = &exchangePartBuffer[3*(nProcs+1)];            //Size nProcs+1
-  int *nreceive    = &exchangePartBuffer[4*(nProcs+1)];            //Size nProcs
-  int *nsendbytes  = &exchangePartBuffer[4*(nProcs+1) + 1*nProcs]; //Size nProcs
-  int *nrecvbytes  = &exchangePartBuffer[4*(nProcs+1) + 2*nProcs]; //Size nProcs
-
-
-
-  memset(nparticles,  0, sizeof(int)*(nProcs+1));
-  memset(nreceive,    0, sizeof(int)*(nProcs));
-  memset(nsendbytes,  0, sizeof(int)*(nProcs));
-  memset(nsendDispls, 0, sizeof(int)*(nProcs));
-
-  // Loop over particles and determine which particle needs to go where
-  // reorder the bodies in such a way that bodies that have to be send
-  // away are stored after each other in the array
-  double t1 = get_time();
-
-  //Array reserve some memory before hand , 1%
-  static vector<bodyStruct> array2Send;
-  array2Send.reserve((int)(nToSend*1.5));
-  array2Send.clear();
-
-
-
-  static int firstSort = 1;
-  if(firstSort)
-  {
-    //This only works if particles are sorted. WHich we currently
-    //only guarentee during the first step
-
-    std::vector<int> offsets(nProcs+2);
-    std::vector<int> items(nProcs+2);
-
-    int location       = 0;
-    offsets[location]  = 0;
-
-    for(int i=0; i < nToSend; i++)
-    {
-      uint4 key  = particlesToSend[i].key;
-
-      bool assigned = false;
-      while(!assigned)
-      {
-        uint4 lowerBoundary = tree.parallelBoundaries[location];
-        uint4 upperBoundary = tree.parallelBoundaries[location+1];
-
-        int bottom = cmp_uint4(key, lowerBoundary);
-        int top    = cmp_uint4(key, upperBoundary);
-
-        assert(bottom >= 0);
-
-        if(top < 0)
-        {
-          //is in box
-          assigned = true;
-        }
-        else
-        {
-          //outside box
-          offsets[++location] = i;
-          assert(location < nProcs);
-        }
-      }//while
-    }//for
-
-    //Fill remaining processes
-    while(location <= nProcs)
-      offsets[++location] = nToSend;
-
-    //Fill items
-    for(int ib=0; ib < nProcs; ib++)
-    {
-      items[ib] = offsets[ib+1]-offsets[ib];
-    }
-
-    assert(items[procId] == 0);
-
-    for(int ib =0; ib < nProcs; ib++)
-    {
-      nparticles[ib] = items[ib];
-      nsendbytes[ib] = nparticles[ib]*sizeof(bodyStruct);
-      nsendDispls[ib] = offsets[ib]*sizeof(bodyStruct);
-    }
-
-      char buff[512];
-      sprintf(buff, "Proc: XXX");
-      for(int i=0; i < nProcs; i++)
-      {
-        //sprintf(buff, "%s [%d, %d] ", buff, nparticles[i], nsendbytes[i]);
-        sprintf(buff, "%s [%d, %d] ", buff, i, nparticles[i]);
-      }
-      LOGF(stderr,"%s \n", buff);
-
-    array2Send.clear();
-    array2Send.insert(array2Send.end(), particlesToSend, particlesToSend+nToSend);
-
-#if 1
-    for(int ib=0; ib < nProcs; ib++)
-    {
-      uint4 lowerBoundary = tree.parallelBoundaries[ib];
-      uint4 upperBoundary = tree.parallelBoundaries[ib+1];
-
-      for(int i= offsets[ib]; i <  offsets[ib+1]; i++)
-      {
-        uint4 key  = particlesToSend[i].key;
-        int bottom = cmp_uint4(key, lowerBoundary);
-        int top    = cmp_uint4(key, upperBoundary);
-
-        if(bottom >= 0 && top < 0)
-        {
-          //Inside
-        }
-        else
-        {
-          assert(!"Particle not in the box");
-        }
-      }//for i
-    }//for ib
-#endif
-
-
-    //LOGF(stderr,  "EXCHANGE reorder iter: %d  took: %lg \tItems: %d Total-n: %d\n",
-    if(iter == 0 && procId == 0)
-      fprintf(stderr,  "EXCHANGE reorder iter: %d  took: %lg \tItems: %d Total-n: %d\n",
-          iter, get_time()-t1, nToSend, tree.n);
-
-
-    firstSort = 0;
-  }
-  else  {
-    //Sort the statistics, causing the processes with which we interact most to be on top
-
-    //Disabled to get this compiled. If we ever enable this code path we should change this
-#if 0
-    std::sort(fullGrpAndLETRequestStatistics, fullGrpAndLETRequestStatistics+nProcs, cmp_uint2_reverse());
-#endif
-
-
-    for(int ib=0;ib<nproc;ib++)
-    {
-      //    int ibox          = (ib+myid)%nproc;
-      int ibox = fullGrpAndLETRequestStatistics[ib].y;
-
-      firstloc[ibox]    = iloc;      //Index of the first particle send to proc: ibox
-      nsendDispls[ibox] = iloc*sizeof(bodyStruct);
-
-      for(int i=iloc; i<nbody;i++)
-      {
-        uint4 lowerBoundary = tree.parallelBoundaries[ibox];
-        uint4 upperBoundary = tree.parallelBoundaries[ibox+1];
-
-        uint4 key  = particlesToSend[i].key;
-        int bottom = cmp_uint4(key, lowerBoundary);
-        int top    = cmp_uint4(key, upperBoundary);
-
-
-        if(bottom >= 0 && top < 0)
-        {
-          //Reorder the particle information
-          tmpp                  = particlesToSend[iloc];
-          particlesToSend[iloc] = particlesToSend[i];
-          particlesToSend[i]    = tmpp;
-
-          //Put the particle in the array of to send particles
-          array2Send.push_back(particlesToSend[iloc]);
-
-          iloc++;
-        }// end if
-      }//for i=iloc
-      nparticles[ibox] = iloc-firstloc[ibox];//Number of particles that has to be send to proc: ibox
-      nsendbytes[ibox] = nparticles[ibox]*sizeof(bodyStruct);
-    } // for(int ib=0;ib<nproc;ib++)
-
-
-    //   printf("Required search time: %lg ,proc: %d found in our own box: %d n: %d  to others: %ld \n",
-    //          get_time()-t1, myid, nparticles[myid], tree.n, array2Send.size());
-    if(iloc < nbody)
-    {
-      LOGF(stderr, "Exchange_particle error: A particle could not be assigned to a box: iloc: %d total: %d \n", iloc,nbody);
-      exit(0);
-    }
-    LOGF(stderr,  "EXCHANGE reorder iter: %d  took: %lg \tItems: %d Total-n: %d\n",
-        iter, get_time()-t1, nToSend, tree.n);
-  }
-
-  double tReorder = get_time();
-
-
-  t1 = get_time();
-#if 0
-  //AlltoallV version
-  MPI_Alltoall(nparticles, 1, MPI_INT, nreceive, 1, MPI_INT, mpiCommWorld);
-
-  //Compute how much we will receive and the offsets and displacements
-  nrecvDispls[0]   = 0;
-  nrecvbytes [0]   = nreceive[0]*sizeof(bodyStruct);
-  unsigned int recvCount  = nreceive[0];
   for(int i=1; i < nProcs; i++)
   {
-    nrecvbytes [i] = nreceive[i]*sizeof(bodyStruct);
-    nrecvDispls[i] = nrecvDispls[i-1] + nrecvbytes [i-1];
-    recvCount     += nreceive[i];
+    rmin.x = std::min(rmin.x, (real)curSysState[i].rmin.x);
+    rmin.y = std::min(rmin.y, (real)curSysState[i].rmin.y);
+    rmin.z = std::min(rmin.z, (real)curSysState[i].rmin.z);
+
+    rmax.x = std::max(rmax.x, (real)curSysState[i].rmax.x);
+    rmax.y = std::max(rmax.y, (real)curSysState[i].rmax.y);
+    rmax.z = std::max(rmax.z, (real)curSysState[i].rmax.z);
+
+    currentRLow[i].x = curSysState[i].rmin.x;
+    currentRLow[i].y = curSysState[i].rmin.y;
+    currentRLow[i].z = curSysState[i].rmin.z;
+    currentRLow[i].w = curSysState[i].rmin.w;
+
+    currentRHigh[i].x = curSysState[i].rmax.x;
+    currentRHigh[i].y = curSysState[i].rmax.y;
+    currentRHigh[i].z = curSysState[i].rmax.z;
+    currentRHigh[i].w = curSysState[i].rmax.w;
+
+    nSamples[i] = curSysState[i].nsample;
   }
-
-  vector<bodyStruct> recv_buffer3(recvCount);
-
-  MPI_Alltoallv(&array2Send[0],   nsendbytes, nsendDispls, MPI_BYTE,
-      &recv_buffer3[0], nrecvbytes, nrecvDispls, MPI_BYTE,
-      mpiCommWorld);
-
-#elif 0
-
-  //Blocking Send/Recv version
-
-  MPI_Alltoall(nparticles, 1, MPI_INT, nreceive, 1, MPI_INT, mpiCommWorld);
-  double t92 = get_time();
-  unsigned int recvCount  = nreceive[0];
-  for (int i = 1; i < nproc; i++)
-  {
-    recvCount     += nreceive[i];
-  }
-  vector<bodyStruct> recv_buffer3(recvCount);
-  double t93=get_time();
-
-  int recvOffset = 0;
-
-  static MPI_Status stat;
-  for (int dist = 1; dist < nproc; dist++)
-  {
-    const int src = (nproc + myid - dist) % nproc;
-    const int dst = (nproc + myid + dist) % nproc;
-    const int scount = nsendbytes[dst];
-    const int rcount = nreceive[src]*sizeof(bodyStruct);
-    if ((myid/dist) & 1)
-    {
-
-      if (scount > 0) MPI_Send(&array2Send[nsendDispls[dst]/sizeof(bodyStruct)], scount, MPI_BYTE, dst, 1, mpiCommWorld);
-      if (rcount > 0) MPI_Recv(&recv_buffer3[recvOffset], rcount, MPI_BYTE   , src, 1, mpiCommWorld, &stat);
-
-      recvOffset +=  nreceive[src];
-    }
-    else
-    {
-      if (rcount > 0) MPI_Recv(&recv_buffer3[recvOffset], rcount, MPI_BYTE   , src, 1, mpiCommWorld, &stat);
-      if (scount > 0) MPI_Send(&array2Send[nsendDispls[dst]/sizeof(bodyStruct)], scount, MPI_BYTE, dst, 1, mpiCommWorld);
-
-      recvOffset +=  nreceive[src];
-    }
-  }
-
-  double t94 = get_time();
-
-#elif 1
-  //Non-blocking send/recv version
-
-  double tStarta2a = get_time();
-  MPI_Alltoall(nparticles, 1, MPI_INT, nreceive, 1, MPI_INT, mpiCommWorld);
-
-  double tEnda2a = get_time();
-  unsigned int recvCount  = nreceive[0];
-
-  for (int i = 1; i < nproc; i++)
-  {
-    recvCount     += nreceive[i];
-  }
-  static std::vector<bodyStruct> recv_buffer3;
-  recv_buffer3.resize(recvCount);
-
-  int recvOffset = 0;
-
-#define NMAXPROC 32768
-  static MPI_Status stat[NMAXPROC];
-  static MPI_Request req[NMAXPROC*2];
-
-  int nreq = 0;
-  for (int dist = 1; dist < nproc; dist++)
-  {
-    const int src    = (nproc + myid - dist) % nproc;
-    const int dst    = (nproc + myid + dist) % nproc;
-    const int scount = nsendbytes[dst];
-    const int rcount = nreceive[src]*sizeof(bodyStruct);
-
-#if 1
-    if (scount > 0) MPI_Isend(&array2Send[nsendDispls[dst]/sizeof(bodyStruct)], scount, MPI_BYTE, dst, 1, mpiCommWorld, &req[nreq++]);
-    if(rcount > 0)
-    {
-      MPI_Irecv(&recv_buffer3[recvOffset], rcount, MPI_BYTE, src, 1, mpiCommWorld, &req[nreq++]);
-      recvOffset += nreceive[src];
-    }
-#else
-    MPI_Status stat;
-    MPI_Sendrecv(&array2Send[nsendDispls[dst]/sizeof(bodyStruct)],
-        scount, MPI_BYTE, dst, 1,
-        &recv_buffer3[recvOffset], rcount, MPI_BYTE, src, 1, mpiCommWorld, &stat);
-    recvOffset += nreceive[src];
-#endif
-  }
-
-  double t94 = get_time();
-  MPI_Waitall(nreq, req, stat);
-
-  double tSendEnd = get_time();
-  LOGF(stderr, "EXCHANGE comm iter: %d  a2asize: %lg data-send: %lg data-wait: %lg \n",
-                iter, tEnda2a-tStarta2a, tEnda2a-t94, tSendEnd -t94);
-
-#else
-  //Build in sendRecv version, note do not enable due to alreeady
-  //moved code
-
-  //Allocate two times the amount of memory of that which we send
-  vector<bodyStruct> recv_buffer3(nbody*2);
-  unsigned int recvCount = 0;
-
-  //Exchange the data with the other processors
-  int ibend = -1;
-  int nsend;
-  int isource = 0;
-  for(int ib=nproc-1;ib>0;ib--)
-  {
-    int ibox = (ib+myid)%nproc; //index to send...
-
-    if (ib == nproc-1)
-    {
-      isource= (myid+1)%nproc;
-    }
-    else
-    {
-      isource = (isource+1)%nproc;
-      if (isource == myid)isource = (isource+1)%nproc;
-    }
-
-
-    if(MP_exchange_particle_with_overflow_check<bodyStruct>(ibox, &array2Send[0],
-          recv_buffer3, firstloc[ibox] - nparticles[myid],
-          nparticles[ibox], isource,
-          nsend, recvCount))
-    {
-      ibend = ibox; //Here we get if exchange failed
-      ib = 0;
-    }//end if mp exchang
-  }//end for all boxes
-
-
-  if(ibend == -1){
-
-  }else{
-    //Something went wrong
-    cerr << "ERROR in exchange_particles_with_overflow_check! \n"; exit(0);
-  }
-#endif
-
-  //If we arrive here all particles have been exchanged, move them to the GPU
-
-  LOGF(stderr,"Required inter-process communication time: %lg ,proc: %d\n", get_time()-t1, myid);
-
-  //Compute the new number of particles:
-  int newN = tree.n + recvCount - nToSend;
-
-  LOGF(stderr, "Exchange, received %d \tSend: %d newN: %d\n", recvCount, nToSend, newN);
-
-#if 1
-  if(iter < 1){ /* jb2404 */
-    fprintf(stderr, "Proc: %d Exchange, received %d \tSend: %d newN: %d\n",
-        procId, recvCount, nToSend, newN);
-  }
-#endif
-
-  execStream->sync();   //make certain that the particle movement on the device
-  //is complete before we resize
-
-  double tSyncGPU = get_time();
-
-  //Allocate 10% extra if we have to allocate, to reduce the total number of
-  //memory allocations
-  int memSize = newN;
-  if(tree.bodies_acc0.get_size() < newN)
-    memSize = newN * MULTI_GPU_MEM_INCREASE;
-
-  LOGF(stderr,"Going to allocate memory for %d particles \n", newN);
-
-  //Have to resize the bodies vector to keep the numbering correct
-  //but do not reduce the size since we need to preserve the particles
-  //in the over sized memory
-  tree.bodies_pos. cresize(memSize + 1, false);
-  tree.bodies_acc0.cresize(memSize,     false);
-  tree.bodies_acc1.cresize(memSize,     false);
-  tree.bodies_vel. cresize(memSize,     false);
-  tree.bodies_time.cresize(memSize,     false);
-  tree.bodies_ids. cresize(memSize + 1, false);
-  tree.bodies_Ppos.cresize(memSize + 1, false);
-  tree.bodies_Pvel.cresize(memSize + 1, false);
-  tree.bodies_key. cresize(memSize + 1, false);
-
-  //This one has to be at least the same size as the number of particles in order to
-  //have enough space to store the other buffers
-  //Can only be resized after we are done since we still have
-  //parts of memory pointing to that buffer (extractList)
-  //Note that we allocate some extra memory to make everything texture/memory aligned
-  tree.generalBuffer1.cresize_nocpy(3*(memSize)*4 + 4096, false);
-
-
-  //Now we have to copy the data in batches in case the generalBuffer1 is not large enough
-  //Amount we can store:
-  int spaceInIntSize    = 3*(memSize)*4;
-  int stepSize          = spaceInIntSize / (sizeof(bodyStruct) / sizeof(int));
-
-  my_dev::dev_mem<bodyStruct>  bodyBuffer(devContext);
-
-  int memOffset1 = bodyBuffer.cmalloc_copy(localTree.generalBuffer1,
-      stepSize, 0);
-
-  //  fprintf(stderr, "Exchange, received %d \tSend: %d newN: %d\tItems that can be insert in one step: %d\n",
-  //      recvCount, nToSend, newN, stepSize);
-//  LOGF(stderr, "Exchange, received %d \tSend: %d newN: %d\tItems that can be insert in one step: %d\n",
-//      recvCount, nToSend, newN, stepSize);
-
-  double tAllocComplete = get_time();
-
-  int insertOffset = 0;
-  for(unsigned int i=0; i < recvCount; i+= stepSize)
-  {
-    int items = min(stepSize, (int)(recvCount-i));
-
-    if(items > 0)
-    {
-      //Copy the data from the MPI receive buffers into the GPU-send buffer
-      memcpy(&bodyBuffer[0], &recv_buffer3[insertOffset], sizeof(bodyStruct)*items);
-
-      bodyBuffer.h2d(items);
-
-      //Start the kernel that puts everything in place
-      insertNewParticlesSFC.set_arg<int>(0,    &nToSend);
-      insertNewParticlesSFC.set_arg<int>(1,    &items);
-      insertNewParticlesSFC.set_arg<int>(2,    &tree.n);
-      insertNewParticlesSFC.set_arg<int>(3,    &insertOffset);
-      insertNewParticlesSFC.set_arg<cl_mem>(4, localTree.bodies_Ppos.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(5, localTree.bodies_Pvel.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(6, localTree.bodies_pos.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(7, localTree.bodies_vel.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(8, localTree.bodies_acc0.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(9, localTree.bodies_acc1.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(10, localTree.bodies_time.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(11, localTree.bodies_ids.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(12, localTree.bodies_key.p());
-      insertNewParticlesSFC.set_arg<cl_mem>(13, bodyBuffer.p());
-      insertNewParticlesSFC.setWork(items, 128);
-      insertNewParticlesSFC.execute(execStream->s());
-    }
-
-    insertOffset += items;
-  }
-
-  //Resize the arrays of the tree
-  tree.setN(newN);
-  reallocateParticleMemory(tree);
-
-  double tEnd = get_time();
-
-  char buff5[1024];
-  sprintf(buff5,"EXCHANGEB-%d: tExSort: %lg tExSend: %lg tExGPUSync: %lg tExGPUAlloc: %lg tExGPUSend: %lg\n",
-                procId, tReorder-tStart, tSendEnd-tStarta2a, tSyncGPU-tSendEnd,
-                tAllocComplete-tSyncGPU, tEnd-tAllocComplete);
-  devContext.writeLogEvent(buff5);
-  sprintf(buff5,"EXCHANGEV-%d: ta2aSize: %lg tISendIRecv: %lg tWaitall: %lg\n",
-                procId, tEnda2a-tStarta2a, t94-tEnda2a, tSendEnd-t94);
-  devContext.writeLogEvent(buff5);
-
-#endif
-
-  return 0;
 }
-
 #endif
 
 
+#if 0
+int octree::recursiveBasedTopLEvelsCheckStart(tree_structure &tree,
+    real4 *treeBuffer,
+    real4 *grpCenter,
+    real4 *grpSize,
+    int startGrp,
+    int endGrp,
+    int &DistanceCheck)
+{
+  //Tree info
+  const int nParticles = host_float_as_int(treeBuffer[0].x);
+  const int nNodes     = host_float_as_int(treeBuffer[0].y);
+
+  //  LOGF(stderr,"Working with %d and %d || %d %d\n", nParticles, nNodes, 1+nParticles+nNodes,nTopLevelTrees );
+
+  real4* treeBoxSizes   = &treeBuffer[1+nParticles];
+  real4* treeBoxCenters = &treeBuffer[1+nParticles+nNodes];
+  real4* treeBoxMoments = &treeBuffer[1+nParticles+2*nNodes];
+
+  const int nodeID = 0;
+  uint4 checkNode = make_uint4(nodeID, startGrp, endGrp, 0);
+
+  int DistanceCheckPP = 0;
+  int maxLevel = recursiveTopLevelCheck(checkNode, treeBoxSizes, treeBoxCenters, treeBoxMoments,
+      grpCenter, grpSize, DistanceCheck, DistanceCheckPP, 0);
 
 
+  //  LOGF(stderr, "Finally Max level found: %d Process : %d \n", maxLevel, ibox)
+  return maxLevel;
+}
+#endif
 
+#if 0
+int octree::recursiveTopLevelCheck(uint4 checkNode, real4* treeBoxSizes, real4* treeBoxCenters, real4* treeBoxMoments,
+    real4* grpCenter, real4* grpSize, int &DistanceCheck, int &DistanceCheckPP, int maxLevel)
+{
+  int nodeID = checkNode.x;
+  int grpID  = checkNode.y;
+  int endGrp = checkNode.z;
 
+  real4 nodeCOM  = treeBoxMoments[nodeID*3];
+  real4 nodeSize = treeBoxSizes  [nodeID];
+  real4 nodeCntr = treeBoxCenters[nodeID];
 
+  //     LOGF(stderr,"Checking node: %d grpID: %d endGrp: %d depth: %d \n",nodeID, grpID, endGrp, maxLevel);
+
+  int res = maxLevel;
+
+  nodeCOM.w = nodeCntr.w;
+  for(int grp=grpID; grp < endGrp; grp++)
+  {
+    real4 grpcntr = grpCenter[grp];
+    real4 grpsize = grpSize[grp];
+
+    bool split = false;
+    {
+      //Compute the distance between the group and the cell
+      float3 dr = make_float3(fabs((float)grpcntr.x - nodeCOM.x) - (float)grpsize.x,
+          fabs((float)grpcntr.y - nodeCOM.y) - (float)grpsize.y,
+          fabs((float)grpcntr.z - nodeCOM.z) - (float)grpsize.z);
+
+      dr.x += fabs(dr.x); dr.x *= 0.5f;
+      dr.y += fabs(dr.y); dr.y *= 0.5f;
+      dr.z += fabs(dr.z); dr.z *= 0.5f;
+
+      //Distance squared, no need to do sqrt since opening criteria has been squared
+      float ds2    = dr.x*dr.x + dr.y*dr.y + dr.z*dr.z;
+
+      if (ds2     <= fabs(nodeCOM.w))           split = true;
+      if (fabs(ds2 - fabs(nodeCOM.w)) < 10e-04) split = true; //Limited precision can result in round of errors. Use this as extra safe guard
+
+      //         LOGF(stderr,"Node: %d grp: %d  split: %d || %f %f\n", nodeID, grp, split, ds2, nodeCOM.w);
+    }
+    //       LOGF(stderr,"Node: %d grp: %d  split: %d Leaf: %d \t %d %f \n",nodeID, grp, split, nodeCntr.w <= 0,(host_float_as_int(nodeSize.w) == 0xFFFFFFFF), nodeCntr.w);
+
+    if(split)
+    {
+      if(host_float_as_int(nodeSize.w) == 0xFFFFFFFF)
+      {
+        //We want to split, but then we go to deep. So we need a full tree-walk
+        return -1;
+      }
+
+      int child, nchild;
+      int childinfo = host_float_as_int(nodeSize.w);
+      bool leaf = nodeCntr.w <= 0;
+
+      if(!leaf)
+      {
+        //Node
+        child    =    childinfo & 0x0FFFFFFF;           //Index to the first child of the node
+        nchild   = (((childinfo & 0xF0000000) >> 28)) ; //The number of children this node has
+
+        //Process node children
+        for(int y=child; y < child+nchild; y++)
+        {
+          //Go one level deeper into the tree
+          //uint4 nextCheck = make_uint4(y, grp, endGrp, 0);
+          uint4 nextCheck = make_uint4(y, grpID, endGrp, 0);
+          int res2 = recursiveTopLevelCheck(nextCheck, treeBoxSizes, treeBoxCenters, treeBoxMoments,
+              grpCenter, grpSize, DistanceCheck, DistanceCheckPP, maxLevel+1);
+          if(res2 < 0) return -1;
+
+          res = max(res,res2); //Return max level reached
+        }
+      }//!leaf
+      else
+      {
+        //It is a leaf, no need to check any other groups
+        return res;
+      }
+      //No need to check further groups since this one already succeeded
+      grp = endGrp;
+    }//Split
+  }
+  return res;
+}
 #endif
